Module Base

This module is the toplevel of the Base library; it's what you get when you write open Base.

The goal of Base is both to be a more complete standard library, with richer APIs, and to be more consistent in its design. For instance, in the standard library some things have modules and others don't; in Base, everything is a module.

Base extends some modules and data structures from the standard library, like Array, Buffer, Bytes, Char, Hashtbl, Int32, Int64, Lazy, List, Map, Nativeint, Printf, Random, Set, String, Sys, and Uchar. One key difference is that Base doesn't use exceptions as much as the standard library and instead makes heavy use of the Result type, as in:

  type ('a, 'b) result =
    | Ok of 'a
    | Error of 'b

Base also adds entirely new modules, most notably:

Comparable, Comparator, and Comparisons in lieu of polymorphic compare.

Container, which provides a consistent interface across container-like data structures (arrays, lists, strings).

Result, Error, and Or_error, supporting the or-error pattern.

module Applicative : sig ... end

module Array : sig ... end

module Avltree : sig ... end

A low-level, mutable AVL tree.

module Backtrace : sig ... end

Module for managing stack backtraces.

module Binary_search : sig ... end

Functions for performing binary searches over ordered sequences given length and get functions.

module Binary_searchable : sig ... end

module Blit : sig ... end

module Bool : sig ... end

Boolean type extended to be enumerable, hashable, sexpable, comparable, and stringable.

module Buffer : sig ... end

Extensible character buffers.

module Bytes : sig ... end

module Char : sig ... end

A type for 8-bit characters.

module Comparable : sig ... end

module Comparator : sig ... end

module Comparisons : sig ... end

Interfaces for infix comparison operators and comparison functions.

module Container : sig ... end

module Dynamic : sig ... end

A value of type 'a Dynamic.t is a dynamically scoped variable of type 'a.

module Either : sig ... end

module Equal : sig ... end

This module defines signatures that are to be included in other signatures to ensure a consistent interface to equal functions. There is a signature (S, S1, S2, S3) for each arity of type. Usage looks like:

module Error : sig ... end

module Exn : sig ... end

Exceptions.

module Field : sig ... end

OCaml record field.

module Float : sig ... end

module Floatable : sig ... end

Module type with float conversion functions.

module Fn : sig ... end

Various combinators for functions.

module Formatter : sig ... end

The Format.formatter type from OCaml's standard library, exported here for convenience and compatibility with other libraries.

module Hash = Ppx_hash_lib.Std.Hash

module Hash_set : sig ... end

module Hashable : sig ... end

module Hasher : sig ... end

Signatures required of types which can be used in [@@deriving hash].

module Hashtbl : sig ... end

module Iarray : sig ... end

module Identifiable : sig ... end

module Indexed_container : sig ... end

module Info : sig ... end

module Int : sig ... end

module Int32 : sig ... end

An int of exactly 32 bits, regardless of the machine.

module Int63 : sig ... end

63-bit integers.

module Int64 : sig ... end

64-bit integers.

module Intable : sig ... end

Functor that adds integer conversion functions to a module.

module Int_math : sig ... end

This module implements derived integer operations (e.g., modulo, rounding to multiples) based on other basic operations.

module Invariant : sig ... end

This module defines signatures that are to be included in other signatures to ensure a consistent interface to invariant-style functions. There is a signature (S, S1, S2, S3) for each arity of type. Usage looks like:

module Dictionary_immutable : sig ... end

module Dictionary_mutable : sig ... end

module Lazy : sig ... end

A value of type 'a Lazy.t is a deferred computation, called a suspension, that has a result of type 'a.

module List : sig ... end

Immutable, singly-linked lists, giving fast access to the front of the list, and slow (i.e., O(n)) access to the back of the list. The comparison functions on lists are lexicographic.

module Map : sig ... end

module Maybe_bound : sig ... end

Used for specifying a bound (either upper or lower) as inclusive, exclusive, or unbounded.

module Modes : sig ... end

module Monad : sig ... end

module Nativeint : sig ... end

Processor-native integers.

module Nothing : sig ... end

An uninhabited type. This is useful when interfaces require that a type be specified, but the implementer knows this type will not be used in their implementation of the interface.

module Obj : sig ... end

Versions of Stdlib.Obj functions that work with modes. Note it is never safe to magic a local value to global.

module Option : sig ... end

The option type indicates whether a meaningful value is present. It is frequently used to represent success or failure, using None for failure. To be more descriptive about why a function failed, see the Or_error module.

module Option_array : sig ... end

'a Option_array.t is a compact representation of 'a option array: it avoids allocating heap objects representing Some x, usually representing them with x instead. It uses a special representation for None that's guaranteed to never collide with any representation of Some x.

module Or_error : sig ... end

Type for tracking errors in an Error.t. This is a specialization of the Result type, where the Error constructor carries an Error.t.

module Or_null : sig ... end

'a or_null, or the unboxed option type, represents a value of type 'a or its absence. Similar to option, it has two constructors, Null and This a.

module Ordered_collection_common : sig ... end

Functions for ordered collections.

module Ordering : sig ... end

Ordering is intended to make code that matches on the result of a comparison more concise and easier to read.

module Poly : sig ... end

A module containing the ad-hoc polymorphic comparison functions. Useful when you want to use polymorphic compare in some small scope of a file within which polymorphic compare has been hidden

module Portability_hacks = Basement.Portability_hacks

module Portable_lazy : sig ... end

A value of type 'a Portable_lazy.t is a deferred computation that has a result of type 'a.

module Pretty_printer : sig ... end

A list of pretty printers for various types, for use in toplevels.

module Printf : sig ... end

Functions for formatted output.

module Linked_queue : sig ... end

This module is a Base-style wrapper around OCaml's standard Queue module.

module Nonempty_list : sig ... end

A 'a t represents a non-empty list, as evidenced by the fact that there is no [] variant. The sexp representation is as a regular list (i.e., the same as the Stable.V3 module below).

module Queue : sig ... end

module Random : sig ... end

Pseudo-random number generation.

module Ref : sig ... end

Module for the type ref, mutable indirection cells r containing a value of type 'a, accessed with !r and set by r := a.

module Result : sig ... end

Result is often used to handle error messages.

module Sequence : sig ... end

A sequence of elements that can be produced one at a time, on demand, normally with no sharing.

module Set : sig ... end

module Sexp : sig ... end

module Sexpable : sig ... end

Provides functors for making modules sexpable when you want the sexp representation of one type to be the same as that for some other isomorphic type.

module Sign : sig ... end

A type for representing the sign of a numeric value.

module Sign_or_nan : sig ... end

An extension to Sign with a Nan constructor, for representing the sign of float-like numeric values.

module Source_code_position : sig ... end

One typically obtains a Source_code_position.t using a [%here] expression, which is implemented by the ppx_here preprocessor.

module Stack : sig ... end

module Staged : sig ... end

A type for making staging explicit in the type of a function.

module String : sig ... end

module Stringable : sig ... end

Provides type-specific conversion functions to and from string.

module Sys : sig ... end

Cross-platform system configuration values.

module T : sig ... end

This module defines various abstract interfaces that are convenient when one needs a module that matches a bare signature with just a type. This sometimes occurs in functor arguments and in interfaces.

module Toplevel_value : sig ... end

This module provides a wrapper that can be used to simulate having toplevel items of non-value layout in modules. For example,

module Type_equal : sig ... end

module Uniform_array : sig ... end

Same semantics as 'a Array.t, except it's guaranteed that the representation array is not tagged with Double_array_tag, the tag for float arrays.

module Unit : sig ... end

Module for the type unit.

module Uchar : sig ... end

module Variant : sig ... end

First-class representative of an individual variant in a variant type, used in [@@deriving variants].

module With_return : sig ... end

with_return f allows for something like the return statement in C within f.

module Word_size : sig ... end

For determining the word size that the program is using.

include module type of struct include T end

module type T = T.T

module type T1__any = T.T1__any

module type T2__any__any = T.T2__any__any

module type T3__any__any__any = T.T3__any__any__any

module type T4__any__any__any__any = T.T4__any__any__any__any

module type T4__any__any__any__value = T.T4__any__any__any__value

module type T3__any__any__value = T.T3__any__any__value

module type T4__any__any__value__any = T.T4__any__any__value__any

module type T4__any__any__value__value = T.T4__any__any__value__value

module type T2__any__value = T.T2__any__value

module type T3__any__value__any = T.T3__any__value__any

module type T4__any__value__any__any = T.T4__any__value__any__any

module type T4__any__value__any__value = T.T4__any__value__any__value

module type T3__any__value__value = T.T3__any__value__value

module type T4__any__value__value__any = T.T4__any__value__value__any

module type T4__any__value__value__value = T.T4__any__value__value__value

module type T1 = T.T1

module type T2__value__any = T.T2__value__any

module type T3__value__any__any = T.T3__value__any__any

module type T4__value__any__any__any = T.T4__value__any__any__any

module type T4__value__any__any__value = T.T4__value__any__any__value

module type T3__value__any__value = T.T3__value__any__value

module type T4__value__any__value__any = T.T4__value__any__value__any

module type T4__value__any__value__value = T.T4__value__any__value__value

module type T2 = T.T2

module type T3__value__value__any = T.T3__value__value__any

module type T4__value__value__any__any = T.T4__value__value__any__any

module type T4__value__value__any__value = T.T4__value__value__any__value

module type T3 = T.T3

module type T4__value__value__value__any = T.T4__value__value__value__any

module type T4 = T.T4

module Int_conversions : sig ... end

module Export : sig ... end

include module type of struct include Export end

type 'a array = 'a Array.t

include sig ... end

val compare_array : 'a. ('a -> 'a -> int) -> 'a Array.t -> 'a Array.t -> int

val compare_array__local : 'a. ('a @ local -> ('a @ local -> int) @ local) -> 'a Array.t @ local -> ('a Array.t @ local -> int) @ local

val equal_array : 'a. ('a -> 'a -> bool) -> 'a Array.t -> 'a Array.t -> bool

val equal_array__local : 'a. ('a @ local -> ('a @ local -> bool) @ local) -> 'a Array.t @ local -> ('a Array.t @ local -> bool) @ local

val globalize_array : 'a. ('a @ local -> 'a) -> 'a Array.t @ local -> 'a Array.t

val sexp_of_array : 'a. ('a -> Sexplib0.Sexp.t) -> 'a Array.t -> Sexplib0.Sexp.t

val sexp_of_array__stack : 'a. ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a Array.t @ local -> Sexplib0.Sexp.t @ local

val array_of_sexp : 'a. (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Array.t

val array_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a Array.t Sexplib0.Sexp_grammar.t @@ portable

type bool = Bool.t

val compare_bool__local : bool @ local -> (bool @ local -> int) @ local

val compare_bool : bool -> bool -> int

val equal_bool__local : bool @ local -> (bool @ local -> bool) @ local

val equal_bool : bool -> bool -> bool

val globalize_bool : bool @ local -> bool

val hash_fold_bool : Ppx_hash_lib.Std.Hash.state -> bool -> Ppx_hash_lib.Std.Hash.state

val hash_bool : bool -> Ppx_hash_lib.Std.Hash.hash_value

val bool_of_sexp : Sexplib0.Sexp.t -> bool

val sexp_of_bool : bool -> Sexplib0.Sexp.t

val sexp_of_bool__stack : bool @ local -> Sexplib0.Sexp.t @ local

val bool_sexp_grammar : bool Sexplib0.Sexp_grammar.t

type char = Char.t

val compare_char__local : char @ local -> (char @ local -> int) @ local

val compare_char : char -> char -> int

val equal_char__local : char @ local -> (char @ local -> bool) @ local

val equal_char : char -> char -> bool

val globalize_char : char @ local -> char

val hash_fold_char : Ppx_hash_lib.Std.Hash.state -> char -> Ppx_hash_lib.Std.Hash.state

val hash_char : char -> Ppx_hash_lib.Std.Hash.hash_value

val char_of_sexp : Sexplib0.Sexp.t -> char

val sexp_of_char : char -> Sexplib0.Sexp.t

val sexp_of_char__stack : char @ local -> Sexplib0.Sexp.t @ local

val char_sexp_grammar : char Sexplib0.Sexp_grammar.t

type exn = Exn.t

val sexp_of_exn : exn -> Sexplib0.Sexp.t

type float = Float.t

val compare_float__local : float @ local -> (float @ local -> int) @ local

val compare_float : float -> float -> int

val equal_float__local : float @ local -> (float @ local -> bool) @ local

val equal_float : float -> float -> bool

val globalize_float : float @ local -> float

val hash_fold_float : Ppx_hash_lib.Std.Hash.state -> float -> Ppx_hash_lib.Std.Hash.state

val hash_float : float -> Ppx_hash_lib.Std.Hash.hash_value

val float_of_sexp : Sexplib0.Sexp.t -> float

val sexp_of_float : float -> Sexplib0.Sexp.t

val sexp_of_float__stack : float @ local -> Sexplib0.Sexp.t @ local

val float_sexp_grammar : float Sexplib0.Sexp_grammar.t

type 'a iarray = 'a Iarray.t

include sig ... end

val compare_iarray : 'a. ('a -> 'a -> int) -> 'a Iarray.t -> 'a Iarray.t -> int

val compare_iarray__local : 'a. ('a @ local -> ('a @ local -> int) @ local) -> 'a Iarray.t @ local -> ('a Iarray.t @ local -> int) @ local

val equal_iarray : 'a. ('a -> 'a -> bool) -> 'a Iarray.t -> 'a Iarray.t -> bool

val equal_iarray__local : 'a. ('a @ local -> ('a @ local -> bool) @ local) -> 'a Iarray.t @ local -> ('a Iarray.t @ local -> bool) @ local

val globalize_iarray : 'a. ('a @ local -> 'a) -> 'a Iarray.t @ local -> 'a Iarray.t

val hash_fold_iarray : 'a. (Ppx_hash_lib.Std.Hash.state -> 'a -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'a Iarray.t -> Ppx_hash_lib.Std.Hash.state

val sexp_of_iarray : ('a -> Sexplib0.Sexp.t) -> 'a Iarray.t -> Sexplib0.Sexp.t

val sexp_of_iarray__stack : ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a Iarray.t @ local -> Sexplib0.Sexp.t @ local

val iarray_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a Iarray.t

val iarray_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a Iarray.t Sexplib0.Sexp_grammar.t @@ portable

type int = Int.t

val compare_int__local : int @ local -> (int @ local -> int) @ local

val compare_int : int -> int -> int

val equal_int__local : int @ local -> (int @ local -> bool) @ local

val equal_int : int -> int -> bool

val globalize_int : int @ local -> int

val hash_fold_int : Ppx_hash_lib.Std.Hash.state -> int -> Ppx_hash_lib.Std.Hash.state

val hash_int : int -> Ppx_hash_lib.Std.Hash.hash_value

val int_of_sexp : Sexplib0.Sexp.t -> int

val sexp_of_int : int -> Sexplib0.Sexp.t

val sexp_of_int__stack : int @ local -> Sexplib0.Sexp.t @ local

val int_sexp_grammar : int Sexplib0.Sexp_grammar.t

type int32 = Int32.t

val compare_int32__local : int32 @ local -> (int32 @ local -> int) @ local

val compare_int32 : int32 -> int32 -> int

val equal_int32__local : int32 @ local -> (int32 @ local -> bool) @ local

val equal_int32 : int32 -> int32 -> bool

val globalize_int32 : int32 @ local -> int32

val hash_fold_int32 : Ppx_hash_lib.Std.Hash.state -> int32 -> Ppx_hash_lib.Std.Hash.state

val hash_int32 : int32 -> Ppx_hash_lib.Std.Hash.hash_value

val int32_of_sexp : Sexplib0.Sexp.t -> int32

val sexp_of_int32 : int32 -> Sexplib0.Sexp.t

val sexp_of_int32__stack : int32 @ local -> Sexplib0.Sexp.t @ local

val int32_sexp_grammar : int32 Sexplib0.Sexp_grammar.t

type int64 = Int64.t

val compare_int64__local : int64 @ local -> (int64 @ local -> int) @ local

val compare_int64 : int64 -> int64 -> int

val equal_int64__local : int64 @ local -> (int64 @ local -> bool) @ local

val equal_int64 : int64 -> int64 -> bool

val globalize_int64 : int64 @ local -> int64

val hash_fold_int64 : Ppx_hash_lib.Std.Hash.state -> int64 -> Ppx_hash_lib.Std.Hash.state

val hash_int64 : int64 -> Ppx_hash_lib.Std.Hash.hash_value

val int64_of_sexp : Sexplib0.Sexp.t -> int64

val sexp_of_int64 : int64 -> Sexplib0.Sexp.t

val sexp_of_int64__stack : int64 @ local -> Sexplib0.Sexp.t @ local

val int64_sexp_grammar : int64 Sexplib0.Sexp_grammar.t

type ('a : value_or_null) list = 'a List.t

val compare_list__local : 'a. ('a @ local -> ('a @ local -> int) @ local) -> 'a list @ local -> ('a list @ local -> int) @ local

val compare_list : 'a. ('a -> 'a -> int) -> 'a list -> 'a list -> int

val equal_list__local : 'a. ('a @ local -> ('a @ local -> bool) @ local) -> 'a list @ local -> ('a list @ local -> bool) @ local

val equal_list : 'a. ('a -> 'a -> bool) -> 'a list -> 'a list -> bool

val globalize_list : 'a. ('a @ local -> 'a) -> 'a list @ local -> 'a list

val hash_fold_list : 'a. (Ppx_hash_lib.Std.Hash.state -> 'a -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'a list -> Ppx_hash_lib.Std.Hash.state

val list_of_sexp : 'a. (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a list

val sexp_of_list : 'a. ('a -> Sexplib0.Sexp.t) -> 'a list -> Sexplib0.Sexp.t

val sexp_of_list__stack : 'a. ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a list @ local -> Sexplib0.Sexp.t @ local

val list_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a list Sexplib0.Sexp_grammar.t

type nativeint = Nativeint.t

val compare_nativeint__local : nativeint @ local -> (nativeint @ local -> int) @ local

val compare_nativeint : nativeint -> nativeint -> int

val equal_nativeint__local : nativeint @ local -> (nativeint @ local -> bool) @ local

val equal_nativeint : nativeint -> nativeint -> bool

val globalize_nativeint : nativeint @ local -> nativeint

val hash_fold_nativeint : Ppx_hash_lib.Std.Hash.state -> nativeint -> Ppx_hash_lib.Std.Hash.state

val hash_nativeint : nativeint -> Ppx_hash_lib.Std.Hash.hash_value

val nativeint_of_sexp : Sexplib0.Sexp.t -> nativeint

val sexp_of_nativeint : nativeint -> Sexplib0.Sexp.t

val sexp_of_nativeint__stack : nativeint @ local -> Sexplib0.Sexp.t @ local

val nativeint_sexp_grammar : nativeint Sexplib0.Sexp_grammar.t

type ('a : value_or_null) option = 'a Option.t

val compare_option__local : 'a. ('a @ local -> ('a @ local -> int) @ local) -> 'a option @ local -> ('a option @ local -> int) @ local

val compare_option : 'a. ('a -> 'a -> int) -> 'a option -> 'a option -> int

val equal_option__local : 'a. ('a @ local -> ('a @ local -> bool) @ local) -> 'a option @ local -> ('a option @ local -> bool) @ local

val equal_option : 'a. ('a -> 'a -> bool) -> 'a option -> 'a option -> bool

val globalize_option : 'a. ('a @ local -> 'a) -> 'a option @ local -> 'a option

val hash_fold_option : 'a. (Ppx_hash_lib.Std.Hash.state -> 'a -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'a option -> Ppx_hash_lib.Std.Hash.state

val option_of_sexp : 'a. (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a option

val sexp_of_option : 'a. ('a -> Sexplib0.Sexp.t) -> 'a option -> Sexplib0.Sexp.t

val sexp_of_option__stack : 'a. ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a option @ local -> Sexplib0.Sexp.t @ local

val option_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a option Sexplib0.Sexp_grammar.t

type nonrec ('a : bits64) option__bits64 = 'a Option.Constructors.t__bits64

and ('a : bits32) option__bits32 = 'a Option.Constructors.t__bits32

and ('a : word) option__word = 'a Option.Constructors.t__word

and ('a : float64) option__float64 = 'a Option.Constructors.t__float64

and ('a : float32) option__float32 = 'a Option.Constructors.t__float32

val compare_option__bits64__local : 'a. (('a : bits64) @ local -> (('a : bits64) @ local -> int) @ local) -> ('a : bits64) option__bits64 @ local -> (('a : bits64) option__bits64 @ local -> int) @ local

val compare_option__bits32__local : 'a. (('a : bits32) @ local -> (('a : bits32) @ local -> int) @ local) -> ('a : bits32) option__bits32 @ local -> (('a : bits32) option__bits32 @ local -> int) @ local

val compare_option__word__local : 'a. (('a : word) @ local -> (('a : word) @ local -> int) @ local) -> ('a : word) option__word @ local -> (('a : word) option__word @ local -> int) @ local

val compare_option__float64__local : 'a. (('a : float64) @ local -> (('a : float64) @ local -> int) @ local) -> ('a : float64) option__float64 @ local -> (('a : float64) option__float64 @ local -> int) @ local

val compare_option__float32__local : 'a. (('a : float32) @ local -> (('a : float32) @ local -> int) @ local) -> ('a : float32) option__float32 @ local -> (('a : float32) option__float32 @ local -> int) @ local

val compare_option__bits64 : 'a. (('a : bits64) -> ('a : bits64) -> int) -> ('a : bits64) option__bits64 -> ('a : bits64) option__bits64 -> int

val compare_option__bits32 : 'a. (('a : bits32) -> ('a : bits32) -> int) -> ('a : bits32) option__bits32 -> ('a : bits32) option__bits32 -> int

val compare_option__word : 'a. (('a : word) -> ('a : word) -> int) -> ('a : word) option__word -> ('a : word) option__word -> int

val compare_option__float64 : 'a. (('a : float64) -> ('a : float64) -> int) -> ('a : float64) option__float64 -> ('a : float64) option__float64 -> int

val compare_option__float32 : 'a. (('a : float32) -> ('a : float32) -> int) -> ('a : float32) option__float32 -> ('a : float32) option__float32 -> int

val equal_option__bits64__local : 'a. (('a : bits64) @ local -> (('a : bits64) @ local -> bool) @ local) -> ('a : bits64) option__bits64 @ local -> (('a : bits64) option__bits64 @ local -> bool) @ local

val equal_option__bits32__local : 'a. (('a : bits32) @ local -> (('a : bits32) @ local -> bool) @ local) -> ('a : bits32) option__bits32 @ local -> (('a : bits32) option__bits32 @ local -> bool) @ local

val equal_option__word__local : 'a. (('a : word) @ local -> (('a : word) @ local -> bool) @ local) -> ('a : word) option__word @ local -> (('a : word) option__word @ local -> bool) @ local

val equal_option__float64__local : 'a. (('a : float64) @ local -> (('a : float64) @ local -> bool) @ local) -> ('a : float64) option__float64 @ local -> (('a : float64) option__float64 @ local -> bool) @ local

val equal_option__float32__local : 'a. (('a : float32) @ local -> (('a : float32) @ local -> bool) @ local) -> ('a : float32) option__float32 @ local -> (('a : float32) option__float32 @ local -> bool) @ local

val equal_option__bits64 : 'a. (('a : bits64) -> ('a : bits64) -> bool) -> ('a : bits64) option__bits64 -> ('a : bits64) option__bits64 -> bool

val equal_option__bits32 : 'a. (('a : bits32) -> ('a : bits32) -> bool) -> ('a : bits32) option__bits32 -> ('a : bits32) option__bits32 -> bool

val equal_option__word : 'a. (('a : word) -> ('a : word) -> bool) -> ('a : word) option__word -> ('a : word) option__word -> bool

val equal_option__float64 : 'a. (('a : float64) -> ('a : float64) -> bool) -> ('a : float64) option__float64 -> ('a : float64) option__float64 -> bool

val equal_option__float32 : 'a. (('a : float32) -> ('a : float32) -> bool) -> ('a : float32) option__float32 -> ('a : float32) option__float32 -> bool

val option_of_sexp__bits64 : 'a. (Sexplib0.Sexp.t -> ('a : bits64)) -> Sexplib0.Sexp.t -> ('a : bits64) option__bits64

val option_of_sexp__bits32 : 'a. (Sexplib0.Sexp.t -> ('a : bits32)) -> Sexplib0.Sexp.t -> ('a : bits32) option__bits32

val option_of_sexp__word : 'a. (Sexplib0.Sexp.t -> ('a : word)) -> Sexplib0.Sexp.t -> ('a : word) option__word

val option_of_sexp__float64 : 'a. (Sexplib0.Sexp.t -> ('a : float64)) -> Sexplib0.Sexp.t -> ('a : float64) option__float64

val option_of_sexp__float32 : 'a. (Sexplib0.Sexp.t -> ('a : float32)) -> Sexplib0.Sexp.t -> ('a : float32) option__float32

val sexp_of_option__bits64 : 'a. (('a : bits64) -> Sexplib0.Sexp.t) -> ('a : bits64) option__bits64 -> Sexplib0.Sexp.t

val sexp_of_option__bits32 : 'a. (('a : bits32) -> Sexplib0.Sexp.t) -> ('a : bits32) option__bits32 -> Sexplib0.Sexp.t

val sexp_of_option__word : 'a. (('a : word) -> Sexplib0.Sexp.t) -> ('a : word) option__word -> Sexplib0.Sexp.t

val sexp_of_option__float64 : 'a. (('a : float64) -> Sexplib0.Sexp.t) -> ('a : float64) option__float64 -> Sexplib0.Sexp.t

val sexp_of_option__float32 : 'a. (('a : float32) -> Sexplib0.Sexp.t) -> ('a : float32) option__float32 -> Sexplib0.Sexp.t

val sexp_of_option__bits64__stack : 'a. (('a : bits64) @ local -> Sexplib0.Sexp.t @ local) -> ('a : bits64) option__bits64 @ local -> Sexplib0.Sexp.t @ local

val sexp_of_option__bits32__stack : 'a. (('a : bits32) @ local -> Sexplib0.Sexp.t @ local) -> ('a : bits32) option__bits32 @ local -> Sexplib0.Sexp.t @ local

val sexp_of_option__word__stack : 'a. (('a : word) @ local -> Sexplib0.Sexp.t @ local) -> ('a : word) option__word @ local -> Sexplib0.Sexp.t @ local

val sexp_of_option__float64__stack : 'a. (('a : float64) @ local -> Sexplib0.Sexp.t @ local) -> ('a : float64) option__float64 @ local -> Sexplib0.Sexp.t @ local

val sexp_of_option__float32__stack : 'a. (('a : float32) @ local -> Sexplib0.Sexp.t @ local) -> ('a : float32) option__float32 @ local -> Sexplib0.Sexp.t @ local

type (('a : value_or_null), ('b : value_or_null)) result = ('a, 'b) Result.t

val compare_result__local : 'a 'b. ('a @ local -> ('a @ local -> int) @ local) -> ('b @ local -> ('b @ local -> int) @ local) -> ('a, 'b) result @ local -> (('a, 'b) result @ local -> int) @ local

val compare_result : 'a 'b. ('a -> 'a -> int) -> ('b -> 'b -> int) -> ('a, 'b) result -> ('a, 'b) result -> int

val equal_result__local : 'a 'b. ('a @ local -> ('a @ local -> bool) @ local) -> ('b @ local -> ('b @ local -> bool) @ local) -> ('a, 'b) result @ local -> (('a, 'b) result @ local -> bool) @ local

val equal_result : 'a 'b. ('a -> 'a -> bool) -> ('b -> 'b -> bool) -> ('a, 'b) result -> ('a, 'b) result -> bool

val globalize_result : 'a 'b. ('a @ local -> 'a) -> ('b @ local -> 'b) -> ('a, 'b) result @ local -> ('a, 'b) result

val hash_fold_result : 'a 'b. (Ppx_hash_lib.Std.Hash.state -> 'a -> Ppx_hash_lib.Std.Hash.state) -> (Ppx_hash_lib.Std.Hash.state -> 'b -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('a, 'b) result -> Ppx_hash_lib.Std.Hash.state

val result_of_sexp : 'a 'b. (Sexplib0.Sexp.t -> 'a) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> ('a, 'b) result

val sexp_of_result : 'a 'b. ('a -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> ('a, 'b) result -> Sexplib0.Sexp.t

val sexp_of_result__stack : 'a 'b. ('a @ local -> Sexplib0.Sexp.t @ local) -> ('b @ local -> Sexplib0.Sexp.t @ local) -> ('a, 'b) result @ local -> Sexplib0.Sexp.t @ local

val result_sexp_grammar : 'a 'b. 'a Sexplib0.Sexp_grammar.t -> 'b Sexplib0.Sexp_grammar.t -> ('a, 'b) result Sexplib0.Sexp_grammar.t

type nonrec (('a : bits64), ('b : value_or_null)) result__bits64 = ('a, 'b) Result.t__bits64

and (('a : bits32), ('b : value_or_null)) result__bits32 = ('a, 'b) Result.t__bits32

and (('a : word), ('b : value_or_null)) result__word = ('a, 'b) Result.t__word

and (('a : float64), ('b : value_or_null)) result__float64 = ('a, 'b) Result.t__float64

and (('a : float32), ('b : value_or_null)) result__float32 = ('a, 'b) Result.t__float32

val compare_result__bits64__local : 'a 'b. (('a : bits64) @ local -> (('a : bits64) @ local -> int) @ local) -> ('b @ local -> ('b @ local -> int) @ local) -> (('a : bits64), 'b) result__bits64 @ local -> ((('a : bits64), 'b) result__bits64 @ local -> int) @ local

val compare_result__bits32__local : 'a 'b. (('a : bits32) @ local -> (('a : bits32) @ local -> int) @ local) -> ('b @ local -> ('b @ local -> int) @ local) -> (('a : bits32), 'b) result__bits32 @ local -> ((('a : bits32), 'b) result__bits32 @ local -> int) @ local

val compare_result__word__local : 'a 'b. (('a : word) @ local -> (('a : word) @ local -> int) @ local) -> ('b @ local -> ('b @ local -> int) @ local) -> (('a : word), 'b) result__word @ local -> ((('a : word), 'b) result__word @ local -> int) @ local

val compare_result__float64__local : 'a 'b. (('a : float64) @ local -> (('a : float64) @ local -> int) @ local) -> ('b @ local -> ('b @ local -> int) @ local) -> (('a : float64), 'b) result__float64 @ local -> ((('a : float64), 'b) result__float64 @ local -> int) @ local

val compare_result__float32__local : 'a 'b. (('a : float32) @ local -> (('a : float32) @ local -> int) @ local) -> ('b @ local -> ('b @ local -> int) @ local) -> (('a : float32), 'b) result__float32 @ local -> ((('a : float32), 'b) result__float32 @ local -> int) @ local

val compare_result__bits64 : 'a 'b. (('a : bits64) -> ('a : bits64) -> int) -> ('b -> 'b -> int) -> (('a : bits64), 'b) result__bits64 -> (('a : bits64), 'b) result__bits64 -> int

val compare_result__bits32 : 'a 'b. (('a : bits32) -> ('a : bits32) -> int) -> ('b -> 'b -> int) -> (('a : bits32), 'b) result__bits32 -> (('a : bits32), 'b) result__bits32 -> int

val compare_result__word : 'a 'b. (('a : word) -> ('a : word) -> int) -> ('b -> 'b -> int) -> (('a : word), 'b) result__word -> (('a : word), 'b) result__word -> int

val compare_result__float64 : 'a 'b. (('a : float64) -> ('a : float64) -> int) -> ('b -> 'b -> int) -> (('a : float64), 'b) result__float64 -> (('a : float64), 'b) result__float64 -> int

val compare_result__float32 : 'a 'b. (('a : float32) -> ('a : float32) -> int) -> ('b -> 'b -> int) -> (('a : float32), 'b) result__float32 -> (('a : float32), 'b) result__float32 -> int

val equal_result__bits64__local : 'a 'b. (('a : bits64) @ local -> (('a : bits64) @ local -> bool) @ local) -> ('b @ local -> ('b @ local -> bool) @ local) -> (('a : bits64), 'b) result__bits64 @ local -> ((('a : bits64), 'b) result__bits64 @ local -> bool) @ local

val equal_result__bits32__local : 'a 'b. (('a : bits32) @ local -> (('a : bits32) @ local -> bool) @ local) -> ('b @ local -> ('b @ local -> bool) @ local) -> (('a : bits32), 'b) result__bits32 @ local -> ((('a : bits32), 'b) result__bits32 @ local -> bool) @ local

val equal_result__word__local : 'a 'b. (('a : word) @ local -> (('a : word) @ local -> bool) @ local) -> ('b @ local -> ('b @ local -> bool) @ local) -> (('a : word), 'b) result__word @ local -> ((('a : word), 'b) result__word @ local -> bool) @ local

val equal_result__float64__local : 'a 'b. (('a : float64) @ local -> (('a : float64) @ local -> bool) @ local) -> ('b @ local -> ('b @ local -> bool) @ local) -> (('a : float64), 'b) result__float64 @ local -> ((('a : float64), 'b) result__float64 @ local -> bool) @ local

val equal_result__float32__local : 'a 'b. (('a : float32) @ local -> (('a : float32) @ local -> bool) @ local) -> ('b @ local -> ('b @ local -> bool) @ local) -> (('a : float32), 'b) result__float32 @ local -> ((('a : float32), 'b) result__float32 @ local -> bool) @ local

val equal_result__bits64 : 'a 'b. (('a : bits64) -> ('a : bits64) -> bool) -> ('b -> 'b -> bool) -> (('a : bits64), 'b) result__bits64 -> (('a : bits64), 'b) result__bits64 -> bool

val equal_result__bits32 : 'a 'b. (('a : bits32) -> ('a : bits32) -> bool) -> ('b -> 'b -> bool) -> (('a : bits32), 'b) result__bits32 -> (('a : bits32), 'b) result__bits32 -> bool

val equal_result__word : 'a 'b. (('a : word) -> ('a : word) -> bool) -> ('b -> 'b -> bool) -> (('a : word), 'b) result__word -> (('a : word), 'b) result__word -> bool

val equal_result__float64 : 'a 'b. (('a : float64) -> ('a : float64) -> bool) -> ('b -> 'b -> bool) -> (('a : float64), 'b) result__float64 -> (('a : float64), 'b) result__float64 -> bool

val equal_result__float32 : 'a 'b. (('a : float32) -> ('a : float32) -> bool) -> ('b -> 'b -> bool) -> (('a : float32), 'b) result__float32 -> (('a : float32), 'b) result__float32 -> bool

val result_of_sexp__bits64 : 'a 'b. (Sexplib0.Sexp.t -> ('a : bits64)) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> (('a : bits64), 'b) result__bits64

val result_of_sexp__bits32 : 'a 'b. (Sexplib0.Sexp.t -> ('a : bits32)) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> (('a : bits32), 'b) result__bits32

val result_of_sexp__word : 'a 'b. (Sexplib0.Sexp.t -> ('a : word)) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> (('a : word), 'b) result__word

val result_of_sexp__float64 : 'a 'b. (Sexplib0.Sexp.t -> ('a : float64)) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> (('a : float64), 'b) result__float64

val result_of_sexp__float32 : 'a 'b. (Sexplib0.Sexp.t -> ('a : float32)) -> (Sexplib0.Sexp.t -> 'b) -> Sexplib0.Sexp.t -> (('a : float32), 'b) result__float32

val sexp_of_result__bits64 : 'a 'b. (('a : bits64) -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> (('a : bits64), 'b) result__bits64 -> Sexplib0.Sexp.t

val sexp_of_result__bits32 : 'a 'b. (('a : bits32) -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> (('a : bits32), 'b) result__bits32 -> Sexplib0.Sexp.t

val sexp_of_result__word : 'a 'b. (('a : word) -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> (('a : word), 'b) result__word -> Sexplib0.Sexp.t

val sexp_of_result__float64 : 'a 'b. (('a : float64) -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> (('a : float64), 'b) result__float64 -> Sexplib0.Sexp.t

val sexp_of_result__float32 : 'a 'b. (('a : float32) -> Sexplib0.Sexp.t) -> ('b -> Sexplib0.Sexp.t) -> (('a : float32), 'b) result__float32 -> Sexplib0.Sexp.t

val sexp_of_result__bits64__stack : 'a 'b. (('a : bits64) @ local -> Sexplib0.Sexp.t @ local) -> ('b @ local -> Sexplib0.Sexp.t @ local) -> (('a : bits64), 'b) result__bits64 @ local -> Sexplib0.Sexp.t @ local

val sexp_of_result__bits32__stack : 'a 'b. (('a : bits32) @ local -> Sexplib0.Sexp.t @ local) -> ('b @ local -> Sexplib0.Sexp.t @ local) -> (('a : bits32), 'b) result__bits32 @ local -> Sexplib0.Sexp.t @ local

val sexp_of_result__word__stack : 'a 'b. (('a : word) @ local -> Sexplib0.Sexp.t @ local) -> ('b @ local -> Sexplib0.Sexp.t @ local) -> (('a : word), 'b) result__word @ local -> Sexplib0.Sexp.t @ local

val sexp_of_result__float64__stack : 'a 'b. (('a : float64) @ local -> Sexplib0.Sexp.t @ local) -> ('b @ local -> Sexplib0.Sexp.t @ local) -> (('a : float64), 'b) result__float64 @ local -> Sexplib0.Sexp.t @ local

val sexp_of_result__float32__stack : 'a 'b. (('a : float32) @ local -> Sexplib0.Sexp.t @ local) -> ('b @ local -> Sexplib0.Sexp.t @ local) -> (('a : float32), 'b) result__float32 @ local -> Sexplib0.Sexp.t @ local

type ('a : value_or_null) ref = 'a Ref.t

val compare_ref__local : 'a. ('a @ local -> ('a @ local -> int) @ local) -> 'a ref @ local -> ('a ref @ local -> int) @ local

val compare_ref : 'a. ('a -> 'a -> int) -> 'a ref -> 'a ref -> int

val equal_ref__local : 'a. ('a @ local -> ('a @ local -> bool) @ local) -> 'a ref @ local -> ('a ref @ local -> bool) @ local

val equal_ref : 'a. ('a -> 'a -> bool) -> 'a ref -> 'a ref -> bool

val globalize_ref : 'a. ('a @ local -> 'a) -> 'a ref @ local -> 'a ref

val ref_of_sexp : 'a. (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a ref

val sexp_of_ref : 'a. ('a -> Sexplib0.Sexp.t) -> 'a ref -> Sexplib0.Sexp.t

val sexp_of_ref__stack : 'a. ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a ref @ local -> Sexplib0.Sexp.t @ local

val ref_sexp_grammar : 'a. 'a Sexplib0.Sexp_grammar.t -> 'a ref Sexplib0.Sexp_grammar.t

type string = String.t

val compare_string__local : string @ local -> (string @ local -> int) @ local

val compare_string : string -> string -> int

val equal_string__local : string @ local -> (string @ local -> bool) @ local

val equal_string : string -> string -> bool

val globalize_string : string @ local -> string

val hash_fold_string : Ppx_hash_lib.Std.Hash.state -> string -> Ppx_hash_lib.Std.Hash.state

val hash_string : string -> Ppx_hash_lib.Std.Hash.hash_value

val string_of_sexp : Sexplib0.Sexp.t -> string

val sexp_of_string : string -> Sexplib0.Sexp.t

val sexp_of_string__stack : string @ local -> Sexplib0.Sexp.t @ local

val string_sexp_grammar : string Sexplib0.Sexp_grammar.t

type bytes = Bytes.t

val compare_bytes__local : bytes @ local -> (bytes @ local -> int) @ local

val compare_bytes : bytes -> bytes -> int

val equal_bytes__local : bytes @ local -> (bytes @ local -> bool) @ local

val equal_bytes : bytes -> bytes -> bool

val globalize_bytes : bytes @ local -> bytes

val bytes_of_sexp : Sexplib0.Sexp.t -> bytes

val sexp_of_bytes : bytes -> Sexplib0.Sexp.t

val sexp_of_bytes__stack : bytes @ local -> Sexplib0.Sexp.t @ local

val bytes_sexp_grammar : bytes Sexplib0.Sexp_grammar.t

type unit = Unit.t

val compare_unit__local : unit @ local -> (unit @ local -> int) @ local

val compare_unit : unit -> unit -> int

val equal_unit__local : unit @ local -> (unit @ local -> bool) @ local

val equal_unit : unit -> unit -> bool

val globalize_unit : unit @ local -> unit

val hash_fold_unit : Ppx_hash_lib.Std.Hash.state -> unit -> Ppx_hash_lib.Std.Hash.state

val hash_unit : unit -> Ppx_hash_lib.Std.Hash.hash_value

val unit_of_sexp : Sexplib0.Sexp.t -> unit

val sexp_of_unit : unit -> Sexplib0.Sexp.t

val sexp_of_unit__stack : unit @ local -> Sexplib0.Sexp.t @ local

val unit_sexp_grammar : unit Sexplib0.Sexp_grammar.t

Format stuff

type nonrec ('a, 'b, 'c) format = ('a, 'b, 'c) format

type nonrec ('a, 'b, 'c, 'd) format4 = ('a, 'b, 'c, 'd) format4

type nonrec ('a, 'b, 'c, 'd, 'e, 'f) format6 = ('a, 'b, 'c, 'd, 'e, 'f) format6

List operators

include module type of struct include List.Infix end

val (@) : 'a. 'a List.t -> 'a List.t -> 'a List.t

Iarray operators

include module type of struct include Iarray.O end

val (.:()) : 'a. 'a Basement.Stdlib_iarray_labels.t -> int -> 'a

An alias for get.

Int operators and comparisons

include module type of struct include Int.O end

val (+) : Int.t -> Int.t -> Int.t

val (-) : Int.t -> Int.t -> Int.t

val (*) : Int.t -> Int.t -> Int.t

val (/) : Int.t -> Int.t -> Int.t

val (~-) : Int.t -> Int.t

val (**) : Int.t @ local -> (Int.t @ local -> Int.t) @ local

val (land) : Int.t -> Int.t -> Int.t

val (lor) : Int.t -> Int.t -> Int.t

val (lxor) : Int.t -> Int.t -> Int.t

val lnot : Int.t @ local -> Int.t

val abs : Int.t -> Int.t

val abs_local : Int.t @ local -> Int.t @ local

val neg : Int.t -> Int.t

val zero : Int.t

val (%) : Int.t @ local -> (Int.t @ local -> Int.t) @ local

val (/%) : Int.t @ local -> (Int.t @ local -> Int.t) @ local

val (//) : Int.t @ local -> (Int.t @ local -> float) @ local

val (lsl) : Int.t -> int -> Int.t

val (asr) : Int.t -> int -> Int.t

val (lsr) : Int.t -> int -> Int.t

val (=) : int -> int -> bool @@ portable

val (<>) : int -> int -> bool @@ portable

val (<) : int -> int -> bool @@ portable

val (>) : int -> int -> bool @@ portable

val (<=) : int -> int -> bool @@ portable

val (>=) : int -> int -> bool @@ portable

val compare : int -> int -> int @@ portable

val compare__local : int -> int -> int @@ portable

val equal : int -> int -> bool @@ portable

val equal__local : int -> int -> bool @@ portable

val ascending : int -> int -> int @@ portable

val descending : int -> int -> int @@ portable

val max : int -> int -> int @@ portable

val min : int -> int -> int @@ portable

Float operators

include module type of struct include Float.O_dot end

val (+.) : Float.t -> Float.t -> Float.t

val (-.) : Float.t -> Float.t -> Float.t

val (*.) : Float.t -> Float.t -> Float.t

val (/.) : Float.t -> Float.t -> Float.t

val (%.) : Float.t -> Float.t -> Float.t

val (**.) : Float.t -> Float.t -> Float.t

val (~-.) : Float.t -> Float.t

val (|>) : 'a 'b. 'a -> ('a -> 'b) -> 'b @@ portable

Reverse application operator. x |> g |> f is equivalent to f (g (x)).

val (@@) : 'a 'b. ('a -> 'b) -> 'a -> 'b @@ portable

Application operator. g @@ f @@ x is equivalent to g (f (x)).

Boolean operations

val (&&) : bool -> bool -> bool @@ portable

val (||) : bool -> bool -> bool @@ portable

val not : bool -> bool @@ portable

val ignore : 'a. 'a @ once immutable -> unit @@ portable

val (^) : String.t @ local -> (String.t @ local -> String.t) @ local

Common string operations

Reference operations

val (!) : 'a. 'a ref -> 'a @@ stateless

val ref : 'a. 'a -> 'a ref @@ stateless

val (:=) : 'a. 'a ref -> 'a -> unit @@ stateless

Pair operations

val fst : 'a 'b. ('a * 'b) -> 'a @@ portable

val snd : 'a 'b. ('a * 'b) -> 'b @@ portable

Exceptions stuff

val raise : 'a. exn -> 'a @ unique portable @@ stateless

val failwith : string -> 'a @ unique portable

val invalid_arg : string -> 'a @ unique portable

val raise_s : Sexplib0.Sexp.t -> 'a @ portable

Misc

val phys_equal : 'a. 'a @ contended -> 'a @ contended -> bool @@ portable

val force : 'a Lazy.t -> 'a @@ portable

type 'a or_null = 'a Or_null.t

val compare_or_null__local : 'a. ('a @ local -> ('a @ local -> int) @ local) -> 'a or_null @ local -> ('a or_null @ local -> int) @ local

val compare_or_null : 'a. ('a -> 'a -> int) -> 'a or_null -> 'a or_null -> int

val equal_or_null__local : 'a. ('a @ local -> ('a @ local -> bool) @ local) -> 'a or_null @ local -> ('a or_null @ local -> bool) @ local

val equal_or_null : 'a. ('a -> 'a -> bool) -> 'a or_null -> 'a or_null -> bool

val globalize_or_null : 'a. ('a @ local -> 'a) -> 'a or_null @ local -> 'a or_null

val hash_fold_or_null : 'a. (Ppx_hash_lib.Std.Hash.state -> 'a -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> 'a or_null -> Ppx_hash_lib.Std.Hash.state

val or_null_of_sexp : 'a. (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a or_null

val sexp_of_or_null : 'a. ('a -> Sexplib0.Sexp.t) -> 'a or_null -> Sexplib0.Sexp.t

val sexp_of_or_null__stack : 'a. ('a @ local -> Sexplib0.Sexp.t @ local) -> 'a or_null @ local -> Sexplib0.Sexp.t @ local

module Continue_or_stop = Container.Export.Continue_or_stop

Continue_or_stop.t is used by the f argument to fold_until in order to indicate whether folding should continue, or stop early.

type ('a : value_or_null) global = 'a Modes.Global.t = {

global : 'a;

}

val compare_global : 'a. (('a : value_or_null) -> ('a : value_or_null) -> int) -> ('a : value_or_null) global -> ('a : value_or_null) global -> int

val compare_global__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> int) @ local) -> ('a : value_or_null) global @ local -> (('a : value_or_null) global @ local -> int) @ local

val equal_global : 'a. (('a : value_or_null) -> ('a : value_or_null) -> bool) -> ('a : value_or_null) global -> ('a : value_or_null) global -> bool

val equal_global__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> bool) @ local) -> ('a : value_or_null) global @ local -> (('a : value_or_null) global @ local -> bool) @ local

val hash_fold_global : 'a. (Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) global -> Ppx_hash_lib.Std.Hash.state

val sexp_of_global : 'a. (('a : value_or_null) -> Sexplib0.Sexp.t) -> ('a : value_or_null) global -> Sexplib0.Sexp.t

val sexp_of_global__stack : 'a. (('a : value_or_null) @ local -> Sexplib0.Sexp.t @ local) -> ('a : value_or_null) global @ local -> Sexplib0.Sexp.t @ local

val global_of_sexp : 'a. (Sexplib0.Sexp.t -> ('a : value_or_null)) -> Sexplib0.Sexp.t -> ('a : value_or_null) global

val global_sexp_grammar : 'a. ('a : value_or_null) Sexplib0.Sexp_grammar.t -> ('a : value_or_null) global Sexplib0.Sexp_grammar.t @@ portable

type ('a : value_or_null) portable = 'a Modes.Portable.t = {

portable : 'a;

}

val compare_portable : 'a. (('a : value_or_null) -> ('a : value_or_null) -> int) -> ('a : value_or_null) portable -> ('a : value_or_null) portable -> int

val compare_portable__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> int) @ local) -> ('a : value_or_null) portable @ local -> (('a : value_or_null) portable @ local -> int) @ local

val equal_portable : 'a. (('a : value_or_null) -> ('a : value_or_null) -> bool) -> ('a : value_or_null) portable -> ('a : value_or_null) portable -> bool

val equal_portable__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> bool) @ local) -> ('a : value_or_null) portable @ local -> (('a : value_or_null) portable @ local -> bool) @ local

val hash_fold_portable : 'a. (Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) portable -> Ppx_hash_lib.Std.Hash.state

val sexp_of_portable : 'a. (('a : value_or_null) -> Sexplib0.Sexp.t) -> ('a : value_or_null) portable -> Sexplib0.Sexp.t

val sexp_of_portable__stack : 'a. (('a : value_or_null) @ local -> Sexplib0.Sexp.t @ local) -> ('a : value_or_null) portable @ local -> Sexplib0.Sexp.t @ local

val portable_sexp_grammar : 'a. ('a : value_or_null) Sexplib0.Sexp_grammar.t -> ('a : value_or_null) portable Sexplib0.Sexp_grammar.t @@ portable

type ('a : value_or_null) contended = 'a Modes.Contended.t = {

contended : 'a;

}

val contended_of_sexp : 'a. (Sexplib0.Sexp.t -> ('a : value_or_null)) -> Sexplib0.Sexp.t -> ('a : value_or_null) contended

type ('a : value_or_null) shared = 'a Modes.Shared.t = {

shared : 'a;

}

val shared_of_sexp : 'a. (Sexplib0.Sexp.t -> ('a : value_or_null)) -> Sexplib0.Sexp.t -> ('a : value_or_null) shared

type ('a : value_or_null) portended = 'a Modes.Portended.t = {

portended : 'a;

}

type ('a : value_or_null) many = 'a Modes.Many.t = {

many : 'a;

}

val compare_many : 'a. (('a : value_or_null) -> ('a : value_or_null) -> int) -> ('a : value_or_null) many -> ('a : value_or_null) many -> int

val compare_many__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> int) @ local) -> ('a : value_or_null) many @ local -> (('a : value_or_null) many @ local -> int) @ local

val equal_many : 'a. (('a : value_or_null) -> ('a : value_or_null) -> bool) -> ('a : value_or_null) many -> ('a : value_or_null) many -> bool

val equal_many__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> bool) @ local) -> ('a : value_or_null) many @ local -> (('a : value_or_null) many @ local -> bool) @ local

val hash_fold_many : 'a. (Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) many -> Ppx_hash_lib.Std.Hash.state

val sexp_of_many : 'a. (('a : value_or_null) -> Sexplib0.Sexp.t) -> ('a : value_or_null) many -> Sexplib0.Sexp.t

val sexp_of_many__stack : 'a. (('a : value_or_null) @ local -> Sexplib0.Sexp.t @ local) -> ('a : value_or_null) many @ local -> Sexplib0.Sexp.t @ local

val many_sexp_grammar : 'a. ('a : value_or_null) Sexplib0.Sexp_grammar.t -> ('a : value_or_null) many Sexplib0.Sexp_grammar.t @@ portable

type ('a : value_or_null) aliased = 'a Modes.Aliased.t = {

aliased : 'a;

}

type ('a : value_or_null) aliased_many = 'a Modes.Aliased_many.t = {

aliased_many : 'a;

}

type ('a : value_or_null) forkable = 'a Modes.Forkable.t = {

forkable : 'a;

}

type ('a : value_or_null) unyielding = 'a Modes.Unyielding.t = {

unyielding : 'a;

}

val compare_unyielding : 'a. (('a : value_or_null) -> ('a : value_or_null) -> int) -> ('a : value_or_null) unyielding -> ('a : value_or_null) unyielding -> int

val compare_unyielding__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> int) @ local) -> ('a : value_or_null) unyielding @ local -> (('a : value_or_null) unyielding @ local -> int) @ local

val equal_unyielding : 'a. (('a : value_or_null) -> ('a : value_or_null) -> bool) -> ('a : value_or_null) unyielding -> ('a : value_or_null) unyielding -> bool

val equal_unyielding__local : 'a. (('a : value_or_null) @ local -> (('a : value_or_null) @ local -> bool) @ local) -> ('a : value_or_null) unyielding @ local -> (('a : value_or_null) unyielding @ local -> bool) @ local

val hash_fold_unyielding : 'a. (Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) -> Ppx_hash_lib.Std.Hash.state) -> Ppx_hash_lib.Std.Hash.state -> ('a : value_or_null) unyielding -> Ppx_hash_lib.Std.Hash.state

val sexp_of_unyielding : 'a. (('a : value_or_null) -> Sexplib0.Sexp.t) -> ('a : value_or_null) unyielding -> Sexplib0.Sexp.t

val sexp_of_unyielding__stack : 'a. (('a : value_or_null) @ local -> Sexplib0.Sexp.t @ local) -> ('a : value_or_null) unyielding @ local -> Sexplib0.Sexp.t @ local

val unyielding_of_sexp : 'a. (Sexplib0.Sexp.t -> ('a : value_or_null)) -> Sexplib0.Sexp.t -> ('a : value_or_null) unyielding

val unyielding_sexp_grammar : 'a. ('a : value_or_null) Sexplib0.Sexp_grammar.t -> ('a : value_or_null) unyielding Sexplib0.Sexp_grammar.t @@ portable

type ('a : value_or_null) stateless = 'a Modes.Stateless.t = {

stateless : 'a;

}

type ('a : value_or_null) observing = 'a Modes.Observing.t = {

observing : 'a;

}

type ('a : value_or_null) immutable = 'a Modes.Immutable.t = {

immutable : 'a;

}

type ('a : value_or_null) read = 'a Modes.Read.t = {

read : 'a;

}

type 'a immutable_data = 'a Modes.Immutable_data.t = {

immutable_data : 'a;

}

exception Not_found_s of Sexplib0.Sexp.t