Module Comparable.Inherit

Inherit comparability from a component.

Parameters

module C : sig ... end

module T : sig ... end

Signature

include Comparisons.S with type t := T.t

include Comparisons.Infix with type t := T.t

val (>=) : T.t -> T.t -> bool

val (<=) : T.t -> T.t -> bool

val (=) : T.t -> T.t -> bool

val (>) : T.t -> T.t -> bool

val (<) : T.t -> T.t -> bool

val (<>) : T.t -> T.t -> bool

val equal : T.t -> T.t -> bool

val compare : T.t -> T.t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : T.t -> T.t -> T.t

val max : T.t -> T.t -> T.t

val ascending : T.t -> T.t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : T.t -> T.t -> int

val between : T.t -> low:T.t -> high:T.t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : T.t -> min:T.t -> max:T.t -> T.t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : T.t -> min:T.t -> max:T.t -> T.t Or_error.t

include Comparator.S with type t := T.t

type comparator_witness

val comparator : (T.t, comparator_witness) Comparator.comparator

val validate_lbound : min:T.t Maybe_bound.t -> T.t Validate.check

val validate_ubound : max:T.t Maybe_bound.t -> T.t Validate.check

val validate_bound : min:T.t Maybe_bound.t -> max:T.t Maybe_bound.t -> T.t Validate.check