Module Container.Continue_or_stop

Continue_or_stop.t is used by the f argument to fold_until in order to indicate whether folding should continue, or stop early.

Continue_or_stop.t is used by the f argument to fold_until in order to indicate whether folding should continue, or stop early.

type ('a, 'b) t =

| Continue of 'a

| Stop of 'b