Module type Dictionary_mutable.Accessors1

type key

type 'data t

val is_empty : _ t -> bool

Whether the dictionary is empty.

val length : _ t -> int

How many key/value pairs the dictionary contains.

val to_alist : 'data t -> (key * 'data) list

All key/value pairs.

val keys : _ t -> key list

All keys in the dictionary, in the same order as to_alist.

val data : 'data t -> 'data list

All values in the dictionary, in the same order as to_alist.

val clear : _ t -> unit

Removes all key/value pairs from the dictionary.

val copy : 'data t -> 'data t

A new dictionary containing the same key/value pairs.

val mem : 'data t -> key -> bool

Whether key has a value.

val find : 'data t -> key -> 'data option

Produces the current value, or absence thereof, for a given key.

val find_exn : 'data t -> key -> 'data

Like find. Raises if there is no value for the given key.

val find_or_add : 'data t -> key -> default:(unit -> 'data) -> 'data

Like find. Adds the value default () if none exists, then returns it.

val findi_or_add : 'data t -> key -> default:(key -> 'data) -> 'data

Like find. Adds default key if no value exists.

val find_and_call : 'data t -> key -> if_found:('data -> 'c) -> if_not_found:(key -> 'c) -> 'c

Like find. Calls if_found data if a value exists, or if_not_found key otherwise. Avoids allocation Some.

val findi_and_call : 'data t -> key -> if_found:(key:key -> data:'data -> 'c) -> if_not_found:(key -> 'c) -> 'c

Like findi. Calls if_found ~key ~data if a value exists.

val find_and_remove : 'data t -> key -> 'data option

Like find. Removes the value for key, if any, from the dictionary before returning it.

val add : 'data t -> key:key -> data:'data -> [ `Ok | `Duplicate ]

Adds a key/value pair for a key the dictionary does not contain, or reports a duplicate.

val add_exn : 'data t -> key:key -> data:'data -> unit

Like add. Raises on duplicates.

val set : 'data t -> key:key -> data:'data -> unit

Adds or replaces a key/value pair in the dictionary.

val remove : 'data t -> key -> unit

Removes any value for the given key.

val change : 'data t -> key -> f:('data option -> 'data option) -> unit

Adds, replaces, or removes the value for a given key, depending on its current value or lack thereof.

val update : 'data t -> key -> f:('data option -> 'data) -> unit

Adds or replaces the value for a given key, depending on its current value or lack thereof.

val update_and_return : 'data t -> key -> f:('data option -> 'data) -> 'data

Like update. Returns the new value.

val incr : ?by:int -> ?remove_if_zero:bool -> int t -> key -> unit

Adds by to the value for key, default 0 if key is absent. May remove key if the result is 0, depending on remove_if_zero.

val decr : ?by:int -> ?remove_if_zero:bool -> int t -> key -> unit

Subtracts by from the value for key, default 0 if key is absent. May remove key if the result is 0, depending on remove_if_zero.

val add_multi : 'data list t -> key:key -> data:'data -> unit

Adds data to the existing key/value pair for key. Interprets a missing key as having an empty list.

val remove_multi : _ list t -> key -> unit

Removes one element from the existing key/value pair for key. Removes the key entirely if the new list is empty.

val find_multi : 'data list t -> key -> 'data list

Produces the list associated with the corresponding key. Interprets a missing key as having an empty list.

val fold : 'data t -> init:'acc -> f:(key:key -> data:'data -> 'acc -> 'acc) -> 'acc

Combines every value in the dictionary.

val for_all : 'data t -> f:('data -> bool) -> bool

Whether every value satisfies f.

val for_alli : 'data t -> f:(key:key -> data:'data -> bool) -> bool

Like for_all. The predicate may also depend on the associated key.

val exists : 'data t -> f:('data -> bool) -> bool

Whether at least one value satisfies f.

val existsi : 'data t -> f:(key:key -> data:'data -> bool) -> bool

Like exists. The predicate may also depend on the associated key.

val count : 'data t -> f:('data -> bool) -> int

How many values satisfy f.

val counti : 'data t -> f:(key:key -> data:'data -> bool) -> int

Like count. The predicate may also depend on the associated key.

val choose : 'data t -> (key * 'data) option

Arbitrary, deterministic key/value pair if non-empty.

val choose_exn : 'data t -> key * 'data

Like choose. Raises if empty.

val choose_randomly : ?random_state:Random.State.t -> 'data t -> (key * 'data) option

Arbitrary, pseudo-random key/value pair if non-empty.

val choose_randomly_exn : ?random_state:Random.State.t -> 'data t -> key * 'data

Like choose_randomly. Raises if empty.

val iter_keys : _ t -> f:(key -> unit) -> unit

Calls f for every key.

val iter : 'data t -> f:('data -> unit) -> unit

Calls f for every value.

val iteri : 'data t -> f:(key:key -> data:'data -> unit) -> unit

Calls f for every key/value pair.

val map : 'data t -> f:('data -> 'c) -> 'c t

Transforms every value.

val mapi : 'data t -> f:(key:key -> data:'data -> 'c) -> 'c t

Like map. The transformation may also depend on the associated key.

val map_inplace : 'data t -> f:('data -> 'data) -> unit

Like map. Modifies the input.

val mapi_inplace : 'data t -> f:(key:key -> data:'data -> 'data) -> unit

Like mapi. Modifies the input.

val filter_keys : 'data t -> f:(key -> bool) -> 'data t

Produces only those key/value pairs whose key satisfies f.

val filter : 'data t -> f:('data -> bool) -> 'data t

Produces only those key/value pairs whose value satisfies f.

val filteri : 'data t -> f:(key:key -> data:'data -> bool) -> 'data t

Produces only those key/value pairs which satisfy f.

val filter_keys_inplace : _ t -> f:(key -> bool) -> unit

Like filter_keys. Modifies the input.

val filter_inplace : 'data t -> f:('data -> bool) -> unit

Like filter. Modifies the input.

val filteri_inplace : 'data t -> f:(key:key -> data:'data -> bool) -> unit

Like filteri. Modifies the input.

val filter_map : 'data t -> f:('data -> 'c option) -> 'c t

Produces key/value pairs for which f produces Some.

val filter_mapi : 'data t -> f:(key:key -> data:'data -> 'c option) -> 'c t

Like filter_map. The new value may also depend on the associated key.

val filter_map_inplace : 'data t -> f:('data -> 'data option) -> unit

Like filter_map. Modifies the input.

val filter_mapi_inplace : 'data t -> f:(key:key -> data:'data -> 'data option) -> unit

Like filter_mapi. Modifies the input.

val partition_tf : 'data t -> f:('data -> bool) -> 'data t * 'data t

Splits one dictionary into two. The first contains key/value pairs for which the value satisfies f. The second contains the remainder.

val partitioni_tf : 'data t -> f:(key:key -> data:'data -> bool) -> 'data t * 'data t

Like partition_tf. The predicate may also depend on the associated key.

val partition_map : 'data t -> f:('data -> ('c, 'd) Either.t) -> 'c t * 'd t

Splits one dictionary into two, corresponding respectively to First _ and Second _ results from f.

val partition_mapi : 'data t -> f:(key:key -> data:'data -> ('c, 'd) Either.t) -> 'c t * 'd t

Like partition_map. The split may also depend on the associated key.

val merge : 'data1 t -> 'data2 t -> f: (key:key -> [ `Left of 'data1 | `Right of 'data2 | `Both of 'data1 * 'data2 ] -> 'data3 option) -> 'data3 t

Merges two dictionaries by fully traversing both. Not suitable for efficiently merging lists of dictionaries. See merge_into instead.

val merge_into : src:'data1 t -> dst:'data2 t -> f:(key:key -> 'data1 -> 'data2 option -> 'data2 Merge_into_action.t) -> unit

Merges two dictionaries by traversing src and adding to dst. Computes the effect on dst of each key/value pair in src using f.