Module Base.Int

type t = int

val globalize : t -> t

include Sexplib0.Sexpable.S with type t := t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

include Floatable.S with type t := t

val of_float : float -> t

val to_float : t -> float

include Intable.S with type t := t

val of_int_exn : int -> t

val to_int_exn : t -> int

include Identifiable.S with type t := t

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Comparable.S with type t := t

include Comparisons.S with type t := t

include Comparisons.Infix with type t := t

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Or_error.t

include Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

include Pretty_printer.S with type t := t

val pp : Formatter.t -> t -> unit

val hashable : t Hashable.t

include Comparable.With_zero with type t := t

val is_positive : t -> bool

val is_non_negative : t -> bool

val is_negative : t -> bool

val is_non_positive : t -> bool

val sign : t -> Sign.t

Returns Neg, Zero, or Pos in a way consistent with the above functions.

val compare__local : t -> t -> int

val equal__local : t -> t -> bool

include Invariant.S with type t := t

val invariant : t -> unit

module Hex : sig ... end

module Binary : sig ... end

val of_string_opt : string -> t option

val to_string_hum : ?delimiter:char -> t -> string

delimiter is an underscore by default.

Infix operators and constants

val one : t

val minus_one : t

Negation

There are two pairs of integer division and remainder functions, /% and %, and / and rem. They both satisfy the same equation relating the quotient and the remainder:

  x = (x /% y) * y + (x % y);
  x = (x /  y) * y + (rem x y);

The functions return the same values if x and y are positive. They all raise if y = 0.

The functions differ if x < 0 or y < 0.

If y < 0, then % and /% raise, whereas / and rem do not.

x % y always returns a value between 0 and y - 1, even when x < 0. On the other hand, rem x y returns a negative value if and only if x < 0; that value satisfies abs (rem x y) <= abs y - 1.

val rem : t -> t -> t

Other common functions

round rounds an int to a multiple of a given to_multiple_of argument, according to a direction dir, with default dir being `Nearest. round will raise if to_multiple_of <= 0. If the result overflows (too far positive or too far negative), round returns an incorrect result.

 | `Down    | rounds toward Int.neg_infinity                          |
 | `Up      | rounds toward Int.infinity                              |
 | `Nearest | rounds to the nearest multiple, or `Up in case of a tie |
 | `Zero    | rounds toward zero                                      |

Here are some examples for round ~to_multiple_of:10 for each direction:

 | `Down    | {10 .. 19} --> 10 | { 0 ... 9} --> 0 | {-10 ... -1} --> -10 |
 | `Up      | { 1 .. 10} --> 10 | {-9 ... 0} --> 0 | {-19 .. -10} --> -10 |
 | `Zero    | {10 .. 19} --> 10 | {-9 ... 9} --> 0 | {-19 .. -10} --> -10 |
 | `Nearest | { 5 .. 14} --> 10 | {-5 ... 4} --> 0 | {-15 ... -6} --> -10 |

For convenience and performance, there are variants of round with dir hard-coded. If you are writing performance-critical code you should use these.

val round : ?dir:[ `Zero | `Nearest | `Up | `Down ] -> t -> to_multiple_of:t -> t

val round_towards_zero : t -> to_multiple_of:t -> t

val round_down : t -> to_multiple_of:t -> t

val round_up : t -> to_multiple_of:t -> t

val round_nearest : t -> to_multiple_of:t -> t

Successor and predecessor functions

val succ : t -> t

val pred : t -> t

Exponentiation

val pow : t -> t -> t

pow base exponent returns base raised to the power of exponent. It is OK if base <= 0. pow raises if exponent < 0, or an integer overflow would occur.

Bit-wise logical operations

val bit_and : t -> t -> t

These are identical to land, lor, etc. except they're not infix and have different names.

val bit_or : t -> t -> t

val bit_xor : t -> t -> t

val bit_not : t -> t

val popcount : t -> int

Returns the number of 1 bits in the binary representation of the input.

Bit-shifting operations

The results are unspecified for negative shifts and shifts >= num_bits.

val shift_left : t -> int -> t

Shifts left, filling in with zeroes.

val shift_right : t -> int -> t

Shifts right, preserving the sign of the input.

Increment and decrement functions for integer references

val decr : t ref -> unit

val incr : t ref -> unit

Conversion functions to related integer types

val of_int32_exn : int32 -> t

val to_int32_exn : t -> int32

val of_int64_exn : int64 -> t

val to_int64 : t -> int64

val of_nativeint_exn : nativeint -> t

val to_nativeint_exn : t -> nativeint

val of_float_unchecked : float -> t

of_float_unchecked truncates the given floating point number to an integer, rounding towards zero. The result is unspecified if the argument is nan or falls outside the range of representable integers.

val num_bits : int

The number of bits available in this integer type. Note that the integer representations are signed.

val max_value : t

The largest representable integer.

val min_value : t

The smallest representable integer.

val shift_right_logical : t -> int -> t

Shifts right, filling in with zeroes, which will not preserve the sign of the input.

val ceil_pow2 : t -> t

ceil_pow2 x returns the smallest power of 2 that is greater than or equal to x. The implementation may only be called for x > 0. Example: ceil_pow2 17 = 32

val floor_pow2 : t -> t

floor_pow2 x returns the largest power of 2 that is less than or equal to x. The implementation may only be called for x > 0. Example: floor_pow2 17 = 16

val ceil_log2 : t -> int

ceil_log2 x returns the ceiling of log-base-2 of x, and raises if x <= 0.

val floor_log2 : t -> int

floor_log2 x returns the floor of log-base-2 of x, and raises if x <= 0.

val is_pow2 : t -> bool

is_pow2 x returns true iff x is a power of 2. is_pow2 raises if x <= 0.

val clz : t -> int

Returns the number of leading zeros in the binary representation of the input, as an integer between 0 and one less than num_bits.

The results are unspecified for t = 0.

val ctz : t -> int

Returns the number of trailing zeros in the binary representation of the input, as an integer between 0 and one less than num_bits.

The results are unspecified for t = 0.

module O : sig ... end

include module type of O

val (+) : t -> t -> t

val (-) : t -> t -> t

val (*) : t -> t -> t

val (/) : t -> t -> t

val (~-) : t -> t

val (**) : t -> t -> t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (>) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (land) : t -> t -> t

val (lor) : t -> t -> t

val (lxor) : t -> t -> t

val lnot : t -> t

val abs : t -> t

val neg : t -> t

val zero : t

val (%) : t -> t -> t

val (/%) : t -> t -> t

val (//) : t -> t -> float

val (lsl) : t -> int -> t

val (asr) : t -> int -> t

val (lsr) : t -> int -> t

val max_value_30_bits : t

max_value_30_bits = 2^30 - 1. It is useful for writing tests that work on both 64-bit and 32-bit platforms.

Conversion functions

val of_int : int -> t

val to_int : t -> int

val of_int32 : int32 -> t option

val to_int32 : t -> int32 option

val of_int64 : int64 -> t option

val of_nativeint : nativeint -> t option

val to_nativeint : t -> nativeint

Truncating conversions

These functions return the least-significant bits of the input. In cases where optional conversions return Some x, truncating conversions return x.

val to_int32_trunc : t -> int32

val of_int32_trunc : int32 -> t

val of_int64_trunc : int64 -> t

val of_nativeint_trunc : nativeint -> t

Byte swap operations

Byte swap operations reverse the order of bytes in an integer. For example, Int32.bswap32 reorders the bottom 32 bits (or 4 bytes), turning 0x1122_3344 to 0x4433_2211. Byte swap functions exposed by Base use OCaml primitives to generate assembly instructions to perform the relevant byte swaps.

For a more extensive list of byteswap functions, see Int32 and Int64.

val bswap16 : int -> int

Byte swaps bottom 16 bits (2 bytes). The values of the remaining bytes are undefined.

Module types specifying integer operations.

module type Binaryable = sig ... end

module type Hexable = sig ... end

module type Int_without_module_types = sig ... end

OCaml's native integer type.

module type Operators = sig ... end

module type Operators_unbounded = sig ... end

module type Round = sig ... end

module type S = sig ... end

module type S_common = sig ... end

module type S_unbounded = sig ... end

module type String_format = sig ... end

module type To_string_format = sig ... end