Module Base.Int32

An int of exactly 32 bits, regardless of the machine.

Side note: There's not much reason to want an int of at least 32 bits (i.e., 32 on 32-bit machines and 63 on 64-bit machines) because Int63 is basically just as efficient.

Overflow issues are not generally considered and explicitly handled. This may be more of an issue for 32-bit ints than 64-bit ints.

Int32.t is boxed on both 32-bit and 64-bit machines.

type t = int32

val globalize : t -> t

include Sexplib0.Sexpable.S with type t := t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

include Floatable.S with type t := t

val of_float : float -> t

val to_float : t -> float

include Intable.S with type t := t

val of_int_exn : int -> t

val to_int_exn : t -> int

include Identifiable.S with type t := t

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

include Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Comparable.S with type t := t

include Comparisons.S with type t := t

include Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Or_error.t

include Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

include Pretty_printer.S with type t := t

val pp : Formatter.t -> t -> unit

val hashable : t Hashable.t

include Comparable.With_zero with type t := t

val is_positive : t -> bool

val is_non_negative : t -> bool

val is_negative : t -> bool

val is_non_positive : t -> bool

val sign : t -> Sign.t

Returns Neg, Zero, or Pos in a way consistent with the above functions.

include Invariant.S with type t := t

val invariant : t -> unit

module Hex : sig ... end

val of_string_opt : string -> t option

val to_string_hum : ?delimiter:char -> t -> string

delimiter is an underscore by default.

Infix operators and constants

val zero : t

val one : t

val minus_one : t

val (+) : t -> t -> t

val (-) : t -> t -> t

val (*) : t -> t -> t

val (**) : t -> t -> t

Integer exponentiation

Negation

val neg : t -> t

val (~-) : t -> t

There are two pairs of integer division and remainder functions, /% and %, and / and rem. They both satisfy the same equation relating the quotient and the remainder:

x = (x /% y) * y + (x % y);
x = (x /  y) * y + (rem x y);

The functions return the same values if x and y are positive. They all raise if y = 0.

The functions differ if x < 0 or y < 0.

If y < 0, then % and /% raise, whereas / and rem do not.

x % y always returns a value between 0 and y - 1, even when x < 0. On the other hand, rem x y returns a negative value if and only if x < 0; that value satisfies abs (rem x y) <= abs y - 1.

val (/%) : t -> t -> t

val (%) : t -> t -> t

val (/) : t -> t -> t

val rem : t -> t -> t

val (//) : t -> t -> float

Float division of integers.

val (land) : t -> t -> t

Same as bit_and.

val (lor) : t -> t -> t

Same as bit_or.

val (lxor) : t -> t -> t

Same as bit_xor.

val lnot : t -> t

Same as bit_not.

val (lsl) : t -> int -> t

Same as shift_left.

val (asr) : t -> int -> t

Same as shift_right.

Other common functions

round rounds an int to a multiple of a given to_multiple_of argument, according to a direction dir, with default dir being `Nearest. round will raise if to_multiple_of <= 0. If the result overflows (too far positive or too far negative), round returns an incorrect result.

       | `Down    | rounds toward Int.neg_infinity                          |
       | `Up      | rounds toward Int.infinity                              |
       | `Nearest | rounds to the nearest multiple, or `Up in case of a tie |
       | `Zero    | rounds toward zero                                      |

Here are some examples for round ~to_multiple_of:10 for each direction:

       | `Down    | {10 .. 19} --> 10 | { 0 ... 9} --> 0 | {-10 ... -1} --> -10 |
       | `Up      | { 1 .. 10} --> 10 | {-9 ... 0} --> 0 | {-19 .. -10} --> -10 |
       | `Zero    | {10 .. 19} --> 10 | {-9 ... 9} --> 0 | {-19 .. -10} --> -10 |
       | `Nearest | { 5 .. 14} --> 10 | {-5 ... 4} --> 0 | {-15 ... -6} --> -10 |

For convenience and performance, there are variants of round with dir hard-coded. If you are writing performance-critical code you should use these.

val round : ?dir:[ `Zero | `Nearest | `Up | `Down ] -> t -> to_multiple_of:t -> t

val round_towards_zero : t -> to_multiple_of:t -> t

val round_down : t -> to_multiple_of:t -> t

val round_up : t -> to_multiple_of:t -> t

val round_nearest : t -> to_multiple_of:t -> t

val abs : t -> t

Returns the absolute value of the argument. May be negative if the input is min_value.

Successor and predecessor functions

val succ : t -> t

val pred : t -> t

Exponentiation

val pow : t -> t -> t

pow base exponent returns base raised to the power of exponent. It is OK if base <= 0. pow raises if exponent < 0, or an integer overflow would occur.

Bit-wise logical operations

val bit_and : t -> t -> t

These are identical to land, lor, etc. except they're not infix and have different names.

val bit_or : t -> t -> t

val bit_xor : t -> t -> t

val bit_not : t -> t

val popcount : t -> int

Returns the number of 1 bits in the binary representation of the input.

Bit-shifting operations

The results are unspecified for negative shifts and shifts >= num_bits.

val shift_left : t -> int -> t

Shifts left, filling in with zeroes.

val shift_right : t -> int -> t

Shifts right, preserving the sign of the input.

Increment and decrement functions for integer references

val decr : t ref -> unit

val incr : t ref -> unit

Conversion functions to related integer types

val of_int32_exn : int32 -> t

val to_int32_exn : t -> int32

val of_int64_exn : int64 -> t

val to_int64 : t -> int64

val of_nativeint_exn : nativeint -> t

val to_nativeint_exn : t -> nativeint

val of_float_unchecked : float -> t

of_float_unchecked truncates the given floating point number to an integer, rounding towards zero. The result is unspecified if the argument is nan or falls outside the range of representable integers.

val num_bits : int

The number of bits available in this integer type. Note that the integer representations are signed.

val max_value : t

The largest representable integer.

val min_value : t

The smallest representable integer.

val (lsr) : t -> int -> t

Same as shift_right_logical.

val shift_right_logical : t -> int -> t

Shifts right, filling in with zeroes, which will not preserve the sign of the input.

val ceil_pow2 : t -> t

ceil_pow2 x returns the smallest power of 2 that is greater than or equal to x. The implementation may only be called for x > 0. Example: ceil_pow2 17 = 32

val floor_pow2 : t -> t

floor_pow2 x returns the largest power of 2 that is less than or equal to x. The implementation may only be called for x > 0. Example: floor_pow2 17 = 16

val ceil_log2 : t -> int

ceil_log2 x returns the ceiling of log-base-2 of x, and raises if x <= 0.

val floor_log2 : t -> int

floor_log2 x returns the floor of log-base-2 of x, and raises if x <= 0.

val is_pow2 : t -> bool

is_pow2 x returns true iff x is a power of 2. is_pow2 raises if x <= 0.

val clz : t -> int

Returns the number of leading zeros in the binary representation of the input, as an integer between 0 and one less than num_bits.

The results are unspecified for t = 0.

val ctz : t -> int

Returns the number of trailing zeros in the binary representation of the input, as an integer between 0 and one less than num_bits.

The results are unspecified for t = 0.

module O : sig ... end

A sub-module designed to be opened to make working with ints more convenient.

Conversion functions

val of_int : int -> t option

val to_int : t -> int option

val of_int32 : int32 -> t

val to_int32 : t -> int32

val of_nativeint : nativeint -> t option

val to_nativeint : t -> nativeint

val of_int64 : int64 -> t option

Truncating conversions

These functions return the least-significant bits of the input. In cases where optional conversions return Some x, truncating conversions return x.

val of_int_trunc : int -> t

val to_int_trunc : t -> int

val of_nativeint_trunc : nativeint -> t

val of_int64_trunc : int64 -> t

Low-level float conversions

val bits_of_float : float -> t

Rounds a regular 64-bit OCaml float to a 32-bit IEEE-754 "single" float, and returns its bit representation. We make no promises about the exact rounding behavior, or what happens in case of over- or underflow.

val float_of_bits : t -> float

Creates a 32-bit IEEE-754 "single" float from the given bits, and converts it to a regular 64-bit OCaml float.

Byte swap operations

See Int's byte swap section for a description of Base's approach to exposing byte swap primitives.

When compiling for 64-bit machines, if signedness of the output value does not matter, use byteswap functions for int64, if possible, for better performance. As of writing, 32-bit byte swap operations on 64-bit machines have extra overhead for moving to 32-bit registers and sign-extending values when returning to 64-bit registers.

The x86 instruction sequence that demonstrates the overhead is in base/bench/bench_int.ml

val bswap16 : t -> t

val bswap32 : t -> t