Module Lazy.T_unforcing

This type offers a serialization function sexp_of_t that won't force its argument. Instead, it will serialize the 'a if it is available, or just use a custom string indicating it is not forced. Note that this is not a round-trippable type, thus the type does not expose of_sexp. To be used in debug code, while tracking a Heisenbug, etc.

type nonrec 'a t = 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t