Module Monad.Of_monad_indexed

The same as S except the monad type has three arguments. The second and third are composed across all computation.

Parameters

module Monad : S_indexed

module M : sig ... end

Signature

val (>>=) : ('a, 'i, 'j) M.t -> ('a -> ('b, 'j, 'k) M.t) -> ('b, 'i, 'k) M.t

val (>>|) : ('a, 'i, 'j) M.t -> ('a -> 'b) -> ('b, 'i, 'j) M.t

module Let_syntax : sig ... end

module Monad_infix : sig ... end

Same as Infix, except the monad type has three arguments. The second and third are composed across all computation.

val bind : ('a, 'i, 'j) M.t -> f:('a -> ('b, 'j, 'k) M.t) -> ('b, 'i, 'k) M.t

val return : 'a -> ('a, 'i, 'i) M.t

val map : ('a, 'i, 'j) M.t -> f:('a -> 'b) -> ('b, 'i, 'j) M.t

val join : (('a, 'j, 'k) M.t, 'i, 'j) M.t -> ('a, 'i, 'k) M.t

val ignore_m : (_, 'i, 'j) M.t -> (unit, 'i, 'j) M.t

val all : ('a, 'i, 'i) M.t list -> ('a list, 'i, 'i) M.t

val all_unit : (unit, 'i, 'i) M.t list -> (unit, 'i, 'i) M.t