Parameter Of_monad_indexed.Monad

The same as S except the monad type has three arguments. The second and third are composed across all computation.

type ('a, 'i, 'j) t

val (>>=) : ('a, 'i, 'j) t -> ('a -> ('b, 'j, 'k) t) -> ('b, 'i, 'k) t

val (>>|) : ('a, 'i, 'j) t -> ('a -> 'b) -> ('b, 'i, 'j) t

module Let_syntax : sig ... end

module Monad_infix : sig ... end

Same as Infix, except the monad type has three arguments. The second and third are composed across all computation.

val bind : ('a, 'i, 'j) t -> f:('a -> ('b, 'j, 'k) t) -> ('b, 'i, 'k) t

val return : 'a -> ('a, 'i, 'i) t

val map : ('a, 'i, 'j) t -> f:('a -> 'b) -> ('b, 'i, 'j) t

val join : (('a, 'j, 'k) t, 'i, 'j) t -> ('a, 'i, 'k) t

val ignore_m : (_, 'i, 'j) t -> (unit, 'i, 'j) t

val all : ('a, 'i, 'i) t list -> ('a list, 'i, 'i) t

val all_unit : (unit, 'i, 'i) t list -> (unit, 'i, 'i) t