Module type Monad.Syntax

Opening a module of this type allows one to use the %bind and %map syntax extensions defined by ppx_let, and brings return into scope.

type 'a t

module Let_syntax : sig ... end

These are convenient to have in scope when programming with a monad: