Module Base.Nativeint

Processor-native integers.

type t = nativeint

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include Sexpable.S with type t := t

val t_of_sexp : Sexp.t -> t

val sexp_of_t : t -> Sexp.t

include Floatable.S with type t := t

val of_float : float -> t

val to_float : t -> float

include Intable.S with type t := t

val of_int_exn : int -> t

val to_int_exn : t -> int

include Identifiable.S with type t := t

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include Sexpable.S with type t := t

val t_of_sexp : Sexp.t -> t

val sexp_of_t : t -> Sexp.t

include Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Comparable.S with type t := t

include Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Or_error.t

include Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

val validate_lbound : min:t Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Maybe_bound.t -> t Validate.check

val validate_bound : min:t Maybe_bound.t -> max:t Maybe_bound.t -> t Validate.check

include Pretty_printer.S with type t := t

val pp : Formatter.t -> t -> unit

include Comparable.With_zero with type t := t

val validate_positive : t Validate.check

val validate_non_negative : t Validate.check

val validate_negative : t Validate.check

val validate_non_positive : t Validate.check

val is_positive : t -> bool

val is_non_negative : t -> bool

val is_negative : t -> bool

val is_non_positive : t -> bool

val sign : t -> Base__Sign0.t

Returns Neg, Zero, or Pos in a way consistent with the above functions.

module Hex : sig ... end

val to_string_hum : ?⁠delimiter:char -> t -> string

delimiter is an underscore by default.

Infix operators and constants

val zero : t

val one : t

val minus_one : t

val (+) : t -> t -> t

val (-) : t -> t -> t

val (*) : t -> t -> t

val (**) : t -> t -> t

Integer exponentiation

val neg : t -> t

val (~-) : t -> t

val (/%) : t -> t -> t

val (%) : t -> t -> t

val (/) : t -> t -> t

val rem : t -> t -> t

val (//) : t -> t -> float

Float division of integers.

val (land) : t -> t -> t

Same as bit_and.

val (lor) : t -> t -> t

Same as bit_or.

val (lxor) : t -> t -> t

Same as bit_xor.

val (lnot) : t -> t

Same as bit_not.

val (lsl) : t -> int -> t

Same as shift_left.

val (asr) : t -> int -> t

Same as shift_right.

Other common functions

val round : ?⁠dir:[ `Zero | `Nearest | `Up | `Down ] -> t -> to_multiple_of:t -> t

val round_towards_zero : t -> to_multiple_of:t -> t

val round_down : t -> to_multiple_of:t -> t

val round_up : t -> to_multiple_of:t -> t

val round_nearest : t -> to_multiple_of:t -> t

val abs : t -> t

Returns the absolute value of the argument. May be negative if the input is min_value.

Successor and predecessor functions

val succ : t -> t

val pred : t -> t

Exponentiation

val pow : t -> t -> t

pow base exponent returns base raised to the power of exponent. It is OK if base <= 0. pow raises if exponent < 0, or an integer overflow would occur.

Bit-wise logical operations

val bit_and : t -> t -> t

These are identical to land, lor, etc. except they're not infix and have different names.

val bit_or : t -> t -> t

val bit_xor : t -> t -> t

val bit_not : t -> t

val popcount : t -> int

Returns the number of 1 bits in the binary representation of the input.

Bit-shifting operations

val shift_left : t -> int -> t

Shifts left, filling in with zeroes.

val shift_right : t -> int -> t

Shifts right, preserving the sign of the input.

Increment and decrement functions for integer references

val decr : t Caml.ref -> unit

val incr : t Caml.ref -> unit

Conversion functions to related integer types

val of_int32_exn : int32 -> t

val to_int32_exn : t -> int32

val of_int64_exn : int64 -> t

val to_int64 : t -> int64

val of_nativeint_exn : nativeint -> t

val to_nativeint_exn : t -> nativeint

val of_float_unchecked : float -> t

of_float_unchecked truncates the given floating point number to an integer, rounding towards zero. The result is unspecified if the argument is nan or falls outside the range of representable integers.

val num_bits : int

The number of bits available in this integer type. Note that the integer representations are signed.

val max_value : t

The largest representable integer.

val min_value : t

The smallest representable integer.

val (lsr) : t -> int -> t

Same as shift_right_logical.

val shift_right_logical : t -> int -> t

Shifts right, filling in with zeroes, which will not preserve the sign of the input.

val ceil_pow2 : t -> t

ceil_pow2 x returns the smallest power of 2 that is greater than or equal to x. The implementation may only be called for x > 0. Example: ceil_pow2 17 = 32

val floor_pow2 : t -> t

floor_pow2 x returns the largest power of 2 that is less than or equal to x. The implementation may only be called for x > 0. Example: floor_pow2 17 = 16

val ceil_log2 : t -> int

ceil_log2 x returns the ceiling of log-base-2 of x, and raises if x <= 0.

val floor_log2 : t -> int

floor_log2 x returns the floor of log-base-2 of x, and raises if x <= 0.

val is_pow2 : t -> bool

is_pow2 x returns true iff x is a power of 2. is_pow2 raises if x <= 0.

module O : sig ... end

A sub-module designed to be opened to make working with ints more convenient.

Conversion functions

val of_int : int -> t

val to_int : t -> int option

val of_int32 : int32 -> t

val to_int32 : t -> int32 option

val of_nativeint : nativeint -> t

val to_nativeint : t -> nativeint

val of_int64 : int64 -> t option

Truncating conversions

val to_int_trunc : t -> int

val to_int32_trunc : t -> int32

val of_int64_trunc : int64 -> t