Module type Queue.S

type 'a t

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

val t_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a t Sexplib0.Sexp_grammar.t

include Indexed_container.S1 with type 'a t := 'a t

include Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'acc -> f:('acc -> 'a -> 'acc) -> 'acc

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'e) Result.t) -> ('acc, 'e) Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'final) Container.Continue_or_stop.t) -> finish:('acc -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Container.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

These are all like their equivalents in Container except that an index starting at 0 is added as the first argument to f.

val foldi : 'a t -> init:_ -> f:(int -> _ -> 'a -> _) -> _

val iteri : 'a t -> f:(int -> 'a -> unit) -> unit

val existsi : 'a t -> f:(int -> 'a -> bool) -> bool

val for_alli : 'a t -> f:(int -> 'a -> bool) -> bool

val counti : 'a t -> f:(int -> 'a -> bool) -> int

val findi : 'a t -> f:(int -> 'a -> bool) -> (int * 'a) option

val find_mapi : 'a t -> f:(int -> 'a -> 'b option) -> 'b option

val singleton : 'a -> 'a t

singleton a returns a queue with one element.

val of_list : 'a list -> 'a t

of_list list returns a queue t with the elements of list in the same order as the elements of list (i.e. the first element of t is the first element of the list).

val of_array : 'a array -> 'a t

val init : int -> f:(int -> 'a) -> 'a t

init n ~f is equivalent to of_list (List.init n ~f).

val enqueue : 'a t -> 'a -> unit

enqueue t a adds a to the end of t.

val enqueue_all : 'a t -> 'a list -> unit

enqueue_all t list adds all elements in list to t in order of list.

val dequeue : 'a t -> 'a option

dequeue t removes and returns the front element of t, if any.

val dequeue_exn : 'a t -> 'a

val dequeue_and_ignore_exn : 'a t -> unit

dequeue_and_ignore_exn t removes the front element of t, or raises if the queue is empty.

val peek : 'a t -> 'a option

peek t returns but does not remove the front element of t, if any.

val peek_exn : 'a t -> 'a

val clear : _ t -> unit

clear t discards all elements from t.

val copy : 'a t -> 'a t

copy t returns a copy of t.

val map : 'a t -> f:('a -> 'b) -> 'b t

val mapi : 'a t -> f:(int -> 'a -> 'b) -> 'b t

val concat_map : 'a t -> f:('a -> 'b list) -> 'b t

Creates a new queue with elements equal to List.concat_map ~f (to_list t).

val concat_mapi : 'a t -> f:(int -> 'a -> 'b list) -> 'b t

val filter_map : 'a t -> f:('a -> 'b option) -> 'b t

filter_map creates a new queue with elements equal to List.filter_map ~f (to_list t).

val filter_mapi : 'a t -> f:(int -> 'a -> 'b option) -> 'b t

val filter : 'a t -> f:('a -> bool) -> 'a t

filter is like filter_map, except with List.filter.

val filteri : 'a t -> f:(int -> 'a -> bool) -> 'a t

val filter_inplace : 'a t -> f:('a -> bool) -> unit

filter_inplace t ~f removes all elements of t that don't satisfy f. If f raises, t is unchanged. This is inplace in that it modifies t; however, it uses space linear in the final length of t.

val filteri_inplace : 'a t -> f:(int -> 'a -> bool) -> unit