Module Base.Set

module type Elt_plain__local = sig ... end

module type Elt_plain = sig ... end

module Without_comparator = Map.Without_comparator

module With_comparator = Map.With_comparator

module With_first_class_module = Map.With_first_class_module

module Merge_to_sequence_element = Sequence.Merge_with_duplicates_element

module type Accessors_generic = sig ... end

module type Transformers_generic = sig ... end

module type Creators_generic = sig ... end

module type Creators_and_accessors_and_transformers_generic = sig ... end

module type S_poly = sig ... end

module type For_deriving = sig ... end

Sets based on Comparator.S.

Creators require a comparator argument to be passed in, whereas accessors use the comparator provided by the input set.

type (!'elt, !'cmp) t

The type of a set. The first type parameter identifies the type of the element, and the second identifies the comparator, which determines the comparison function that is used for ordering elements in this set. Many operations (e.g., union), require that they be passed sets with the same element type and the same comparator type.

include Ppx_compare_lib.Comparable.S2 with type (!'elt, !'cmp) t := ('elt, 'cmp) t

val compare : ('a -> 'a -> int) -> ('b -> 'b -> int) -> ('a, 'b) t -> ('a, 'b) t -> int

include Ppx_compare_lib.Comparable.S2__local with type (!'elt, !'cmp) t := ('elt, 'cmp) t

val globalize : 'elt 'cmp. ('elt @ local -> 'elt) -> ('cmp @ local -> 'cmp) -> ('elt, 'cmp) t @ local -> ('elt, 'cmp) t

val invariants : (_, _) t -> bool @@ portable

Tests internal invariants of the set data structure. Returns true on success.

val comparator_s : ('a, 'cmp) t -> ('a, 'cmp) Comparator.Module.t @@ portable

Returns a first-class module that can be used to build other map/set/etc with the same notion of comparison.

val comparator : ('a, 'cmp) t -> ('a, 'cmp) Comparator.t @@ portable

val globalize0 : ('a, 'cmp) t @ local -> ('a, 'cmp) t @@ portable

val empty : 'a 'cmp. ('a, 'cmp) Comparator.Module.t -> ('a, 'cmp) t @@ portable

Creates an empty set based on the provided comparator.

val singleton : ('a, 'cmp) Comparator.Module.t -> 'a -> ('a, 'cmp) t @@ portable

Creates a set based on the provided comparator that contains only the provided element.

val length : (_, _) t @ local -> int @@ portable

Returns the cardinality of the set. O(1).

val is_empty : (_, _) t @ local -> bool @@ portable

is_empty t is true iff t is empty. O(1).

val mem : ('a, _) t -> 'a -> bool @@ portable

mem t a returns true iff a is in t. O(log n).

val add : ('a, 'cmp) t -> 'a -> ('a, 'cmp) t @@ portable

add t a returns a new set with a added to t, or returns t if mem t a. O(log n).

val remove : ('a, 'cmp) t -> 'a -> ('a, 'cmp) t @@ portable

remove t a returns a new set with a removed from t if mem t a, or returns t otherwise. O(log n).

val union : ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t @@ portable

union t1 t2 returns the union of the two sets. O(m * log (n/m + 1)), where m is the length of the smaller set and n is the length of the larger set.

val union_list : ('a, 'cmp) Comparator.Module.t -> ('a, 'cmp) t list -> ('a, 'cmp) t @@ portable

union_list c list returns the union of all the sets in list. The comparator argument is required for the case where list is empty. O(max(List.length list, n log n)), where n is the sum of sizes of the input sets.

val inter : ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t @@ portable

inter t1 t2 computes the intersection of sets t1 and t2. O(m * log (n/m + 1)), where m is the length of the smaller set and n is the length of the larger set.

val diff : ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t @@ portable

diff t1 t2 computes the set difference t1 - t2, i.e., the set containing all elements in t1 that are not in t2. O(m * log (n/m + 1)), where m is the length of the smaller set and n is the length of the larger set. .

val symmetric_diff : ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'a) Either.t Sequence.t @@ portable

symmetric_diff t1 t2 returns a sequence of changes between t1 and t2. It is intended to be efficient in the case where t1 and t2 share a large amount of structure.

val compare_direct : ('a, 'cmp) t -> ('a, 'cmp) t -> int @@ portable

compare_direct t1 t2 compares the sets t1 and t2. It returns the same result as compare, but unlike compare, doesn't require arguments to be passed in for the type parameters of the set. O(length t1 + length t2).

val hash_fold_direct : 'a Ppx_hash_lib.Std.Hash.folder -> ('a, 'cmp) t Ppx_hash_lib.Std.Hash.folder @@ portable

Hash function: a building block to use when hashing data structures containing sets in them. hash_fold_direct hash_fold_key is compatible with compare_direct iff hash_fold_key is compatible with (comparator s).compare of the set s being hashed.

val equal : ('a, 'cmp) t -> ('a, 'cmp) t -> bool @@ portable

equal t1 t2 returns true iff the two sets have the same elements. O(length t1 + length t2)

val exists : ('a, _) t -> f:('a -> bool) @ local -> bool @@ portable

exists t ~f returns true iff there exists an a in t for which f a. O(n), but returns as soon as it finds an a for which f a.

val for_all : ('a, _) t -> f:('a -> bool) @ local -> bool @@ portable

for_all t ~f returns true iff for all a in t, f a. O(n), but returns as soon as it finds an a for which not (f a).

val count : ('a, _) t -> f:('a -> bool) @ local -> int @@ portable

count t returns the number of elements of t for which f returns true. O(n).

val sum : (module Container.Summable with type t = 'sum) -> ('a, _) t -> f:('a -> 'sum) @ local -> 'sum @@ portable

sum t returns the sum of f t for each t in the set. O(n).

val find : ('a, _) t -> f:('a -> bool) @ local -> 'a option @@ portable

find t f returns an element of t for which f returns true, with no guarantee as to which element is returned. O(n), but returns as soon as a suitable element is found.

val find_map : ('a, _) t -> f:('a -> 'b option) @ local -> 'b option @@ portable

find_map t f returns b for some a in t for which f a = Some b. If no such a exists, then find returns None. O(n), but returns as soon as a suitable element is found.

val find_exn : ('a, _) t -> f:('a -> bool) @ local -> 'a @@ portable

Like find, but throws an exception on failure.

val nth : ('a, _) t -> int -> 'a option @@ portable

nth t i returns the ith smallest element of t, in O(log n) time. The smallest element has i = 0. Returns None if i < 0 or i >= length t.

val rank : ('a, _) t -> 'a -> int option @@ portable

rank t elt If elt is in t, returns the number of elements strictly less than elt in t, and None otherwise.

val remove_index : ('a, 'cmp) t -> int -> ('a, 'cmp) t @@ portable

remove_index t i returns a version of t with the ith smallest element removed, in O(log n) time. The smallest element has i = 0. Returns t if i < 0 or i >= length t.

val is_subset : ('a, 'cmp) t -> of_:('a, 'cmp) t -> bool @@ portable

is_subset t1 ~of_:t2 returns true iff t1 is a subset of t2.

val are_disjoint : ('a, 'cmp) t -> ('a, 'cmp) t -> bool @@ portable

are_disjoint t1 t2 returns true iff is_empty (inter t1 t2), but is more efficient. O(m * log (n/m + 1)), where m is the length of the smaller set and n is the length of the larger set.

module Named : sig ... end

Named allows the validation of subset and equality relationships between sets. A Named.t is a record of a set and a name, where the name is used in error messages, and Named.is_subset and Named.equal validate subset and equality relationships respectively.

val of_list : ('a, 'cmp) Comparator.Module.t -> 'a list -> ('a, 'cmp) t @@ portable

The list or array given to of_list and of_array need not be sorted.

val of_sequence : ('a, 'cmp) Comparator.Module.t -> 'a Sequence.t -> ('a, 'cmp) t @@ portable

val of_array : ('a, 'cmp) Comparator.Module.t -> 'a array -> ('a, 'cmp) t @@ portable

val to_list : ('a, _) t -> 'a list @@ portable

to_list and to_array produce sequences sorted in ascending order according to the comparator.

val to_array : ('a, _) t -> 'a array @@ portable

val of_sorted_array : ('a, 'cmp) Comparator.Module.t -> 'a array -> ('a, 'cmp) t Or_error.t @@ portable

Create set from sorted array. The input must be sorted (either in ascending or descending order as given by the comparator) and contain no duplicates, otherwise the result is an error. The complexity of this function is O(n).

val of_sorted_array_unchecked : ('a, 'cmp) Comparator.Module.t -> 'a array -> ('a, 'cmp) t @@ portable

Similar to of_sorted_array, but without checking the input array.

val of_increasing_iterator_unchecked : ('a, 'cmp) Comparator.Module.t -> len:int -> f:(int -> 'a) @ local -> ('a, 'cmp) t @@ portable

of_increasing_iterator_unchecked c ~len ~f behaves like of_sorted_array_unchecked c (Array.init len ~f), with the additional restriction that a decreasing order is not supported. The advantage is not requiring you to allocate an intermediate array. f will be called with 0, 1, ... len - 1, in order.

val map : ('b, 'cmp) Comparator.Module.t -> ('a, _) t -> f:('a -> 'b) @ local -> ('b, 'cmp) t @@ portable

map c t ~f returns a new set created by applying f to every element in t. The returned set is based on the provided comparator. O(n log n).

val filter_map : ('b, 'cmp) Comparator.Module.t -> ('a, _) t -> f:('a -> 'b option) @ local -> ('b, 'cmp) t @@ portable

Like map, except elements for which f returns None will be dropped.

val filter : ('a, 'cmp) t -> f:('a -> bool) @ local -> ('a, 'cmp) t @@ portable

filter t ~f returns the subset of t for which f evaluates to true. O(n log n).

val fold : ('a, _) t -> init:'acc -> f:('acc -> ('a -> 'acc) @ local) @ local -> 'acc @@ portable

fold t ~init ~f folds over the elements of the set from smallest to largest.

val fold_result : ('a, _) t -> init:'acc -> f:('acc -> ('a -> ('acc, 'e) Result.t) @ local) @ local -> ('acc, 'e) Result.t @@ portable

fold_result ~init ~f folds over the elements of the set from smallest to largest, short circuiting the fold if f accum x is an Error _

val fold_until : ('a, _) t -> init:'acc -> f: ('acc -> ('a -> ('acc, 'final) Container.Continue_or_stop.t) @ local) @ local -> (finish:('acc -> 'final) @ local -> 'final) @ local @@ portable

fold_until t ~init ~f is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed.

val fold_right : ('a, _) t -> init:'acc -> f:('a -> ('acc -> 'acc) @ local) @ local -> 'acc @@ portable

Like fold, except that it goes from the largest to the smallest element.

val iter : ('a, _) t -> f:('a -> unit) @ local -> unit @@ portable

iter t ~f calls f on every element of t, going in order from the smallest to largest.

val iter_until : ('a, _) t -> f:('a -> (unit, 'final) Container.Continue_or_stop.t) @ local -> (finish:(unit -> 'final) @ local -> 'final) @ local @@ portable

iter_until t ~f ~finish is a short-circuiting version of iter. If f returns Stop x the computation ceases and returns x. If f always returns Continue () the final result is computed by finish.

val iter2 : ('a, 'cmp) t -> ('a, 'cmp) t -> f:([ `Left of 'a | `Right of 'a | `Both of 'a * 'a ] -> unit) @ local -> unit @@ portable

Iterate two sets side by side. Complexity is O(m+n) where m and n are the sizes of the two input sets. As an example, with the inputs 0; 1 and 1; 2, f will be called with `Left 0; `Both (1, 1); and `Right 2.

val partition_tf : ('a, 'cmp) t -> f:('a -> bool) @ local -> ('a, 'cmp) t * ('a, 'cmp) t @@ portable

if a, b = partition_tf set ~f then a is the elements on which f produced true, and b is the elements on which f produces false.

val elements : ('a, _) t -> 'a list @@ portable

Same as to_list.

val min_elt : ('a, _) t -> 'a option @@ portable

Returns the smallest element of the set. O(log n).

val min_elt_exn : ('a, _) t -> 'a @@ portable

Like min_elt, but throws an exception when given an empty set.

val max_elt : ('a, _) t -> 'a option @@ portable

Returns the largest element of the set. O(log n).

val max_elt_exn : ('a, _) t -> 'a @@ portable

Like max_elt, but throws an exception when given an empty set.

val choose : ('a, _) t -> 'a option @@ portable

returns an arbitrary element, or None if the set is empty.

val choose_exn : ('a, _) t -> 'a @@ portable

Like choose, but throws an exception on an empty set.

val split : ('a, 'cmp) t -> 'a -> ('a, 'cmp) t * 'a option * ('a, 'cmp) t @@ portable

split t x produces a triple (t1, maybe_x, t2).

t1 is the set of elements strictly less than x, maybe_x is the member (if any) of t which compares equal to x, t2 is the set of elements strictly larger than x.

val split_le_gt : ('a, 'cmp) t -> 'a -> ('a, 'cmp) t * ('a, 'cmp) t @@ portable

split_le_gt t x produces a pair (t1, t2).

t1 is the set of elements less than or equal to x, t2 is the set of elements strictly greater than x.

val split_lt_ge : ('a, 'cmp) t -> 'a -> ('a, 'cmp) t * ('a, 'cmp) t @@ portable

split_lt_ge t x produces a pair (t1, t2).

t1 is the set of elements strictly less than x, t2 is the set of elements greater than or equal to x.

val group_by : ('a, 'cmp) t -> equiv:('a -> ('a -> bool) @ local) @ local -> ('a, 'cmp) t list @@ portable

if equiv is an equivalence predicate, then group_by set ~equiv produces a list of equivalence classes (i.e., a set-theoretic quotient). E.g.,

  let chars = Set.of_list [ 'A'; 'a'; 'b'; 'c' ] in
  let equiv c c' = Char.equal (Char.uppercase c) (Char.uppercase c') in
  group_by chars ~equiv

produces:

  [ Set.of_list [ 'A'; 'a' ]; Set.singleton 'b'; Set.singleton 'c' ]

group_by runs in O(n^2) time, so if you have a comparison function, it's usually much faster to use Set.of_list.

deprecated [since 2024-08] This function is slow (O(n^2)) and pretty much never the right thing to use. Consider using [to_list] along with [List.sort_and_group] or [List.group].

val to_sequence : ?order:[ `Increasing | `Decreasing ] -> ?greater_or_equal_to:'a -> ?less_or_equal_to:'a -> ('a, 'cmp) t -> 'a Sequence.t @@ portable

to_sequence t converts the set t to a sequence of the elements between greater_or_equal_to and less_or_equal_to inclusive in the order indicated by order. If greater_or_equal_to > less_or_equal_to the sequence is empty. Cost is O(log n) up front and amortized O(1) for each element produced.

val binary_search : ('a, 'cmp) t -> compare:('a -> ('key -> int) @ local) @ local -> ([ `Last_strictly_less_than | `Last_less_than_or_equal_to | `Last_equal_to | `First_equal_to | `First_greater_than_or_equal_to | `First_strictly_greater_than ] -> ('key -> 'a option) @ local) @ local @@ portable

binary_search t ~compare which elt returns the element in t specified by compare and which, if one exists.

t must be sorted in increasing order according to compare, where compare and elt divide t into three (possibly empty) segments:

  |  < elt  |  = elt  |  > elt  |

binary_search returns an element on the boundary of segments as specified by which. See the diagram below next to the which variants.

binary_search does not check that compare orders t, and behavior is unspecified if compare doesn't order t. Behavior is also unspecified if compare mutates t.

val binary_search_segmented : ('a, 'cmp) t -> segment_of:('a -> [ `Left | `Right ]) @ local -> ([ `Last_on_left | `First_on_right ] -> 'a option) @ local @@ portable

binary_search_segmented t ~segment_of which takes a segment_of function that divides t into two (possibly empty) segments:

  | segment_of elt = `Left | segment_of elt = `Right |

binary_search_segmented returns the element on the boundary of the segments as specified by which: `Last_on_left yields the last element of the left segment, while `First_on_right yields the first element of the right segment. It returns None if the segment is empty.

binary_search_segmented does not check that segment_of segments t as in the diagram, and behavior is unspecified if segment_of doesn't segment t. Behavior is also unspecified if segment_of mutates t.

val merge_to_sequence : ?order:[ `Increasing | `Decreasing ] -> ?greater_or_equal_to:'a -> ?less_or_equal_to:'a -> ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'a) Merge_to_sequence_element.t Sequence.t @@ portable

Produces the elements of the two sets between greater_or_equal_to and less_or_equal_to in order, noting whether each element appears in the left set, the right set, or both. In the both case, both elements are returned, in case the caller can distinguish between elements that are equal to the sets' comparator. Runs in O(length t + length t').

module M (Elt : sig ... end) : sig ... end

M is meant to be used in combination with OCaml applicative functor types:

include For_deriving with type ('a, 'b) t := ('a, 'b) t

module type Sexp_of_m = Sexpable.Sexp_of

module type M_of_sexp = sig ... end

module type M_sexp_grammar = sig ... end

module type Compare_m = sig ... end

module type Equal_m = sig ... end

module type Globalize_m = sig ... end

module type Hash_fold_m = Hasher.S

val sexp_of_m__t : (module Sexp_of_m with type t = 'elt) -> ('elt, 'cmp) t -> Sexplib0.Sexp.t

val m__t_of_sexp : (module M_of_sexp with type comparator_witness = 'cmp and type t = 'elt) -> Sexplib0.Sexp.t -> ('elt, 'cmp) t

val m__t_sexp_grammar : (module M_sexp_grammar with type t = 'elt) -> ('elt, 'cmp) t Sexplib0.Sexp_grammar.t @@ portable

val compare_m__t : (module Compare_m) -> ('elt, 'cmp) t -> ('elt, 'cmp) t -> int

val equal_m__t : (module Equal_m) -> ('elt, 'cmp) t -> ('elt, 'cmp) t -> bool

val compare_m__t__local : (module Compare_m) -> ('elt, 'cmp) t @ local -> (('elt, 'cmp) t @ local -> int) @ local

val equal_m__t__local : (module Equal_m) -> ('elt, 'cmp) t @ local -> (('elt, 'cmp) t @ local -> bool) @ local

val globalize_m__t : (module Globalize_m) -> ('elt, 'cmp) t @ local -> ('elt, 'cmp) t

val hash_fold_m__t : (module Hash_fold_m with type t = 'elt) -> Ppx_hash_lib.Std.Hash.state -> ('elt, _) t -> Ppx_hash_lib.Std.Hash.state

val hash_m__t : (module Hash_fold_m with type t = 'elt) -> ('elt, _) t -> int

module Tree : sig ... end

module Using_comparator : sig ... end

Using comparator is a similar interface as the toplevel of Set, except the functions take a ~comparator:('elt, 'cmp) Comparator.t where the functions at the toplevel of Set takes a ('elt, 'cmp) comparator.

val to_tree : ('a, 'cmp) t -> ('a, 'cmp) Using_comparator.Tree.t @@ portable

val of_tree : ('a, 'cmp) Comparator.Module.t -> ('a, 'cmp) Using_comparator.Tree.t -> ('a, 'cmp) t @@ portable

module Poly : S_poly with type 'elt t = ('elt, Comparator.Poly.comparator_witness) t with type comparator_witness := Comparator.Poly.comparator_witness with type 'elt tree := ('elt, Comparator.Poly.comparator_witness) Using_comparator.Tree.t with type ('elt, 'cmp) set := ('elt, 'cmp) t

A polymorphic Set.