Module type Set.Accessors2_with_comparator

include Container.S1_phantom_invariant

type ('a, 'phantom) t

val length : (_, _) t -> int

val is_empty : (_, _) t -> bool

val iter : ('a, _) t -> f:('a -> unit) -> unit

val fold : ('a, _) t -> init:'accum -> f:('accum -> 'a -> 'accum) -> 'accum

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t.

val fold_result : ('a, _) t -> init:'accum -> f:('accum -> 'a -> ('accum, 'e) Result.t) -> ('accum, 'e) Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val exists : ('a, _) t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : ('a, _) t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : ('a, _) t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Base__Container_intf.Summable with type t = 'sum) -> ('a, _) t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : ('a, _) t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : ('a, _) t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : ('a, _) t -> 'a list

val to_array : ('a, _) t -> 'a array

type ('a, 'cmp) tree

val invariants : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> bool

val mem : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> 'a -> bool

val add : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> 'a -> ('a, 'cmp) t

val remove : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> 'a -> ('a, 'cmp) t

val union : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t

val inter : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t

val diff : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'cmp) t

val symmetric_diff : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'a) Either.t Sequence.t

val compare_direct : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> ('a, 'cmp) t -> int

val equal : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> ('a, 'cmp) t -> bool

val is_subset : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> of_:('a, 'cmp) t -> bool

val are_disjoint : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> ('a, 'cmp) t -> bool

type ('a, 'cmp) named

module Named : sig ... end

val fold_until : ('a, _) t -> init:'accum -> f:('accum -> 'a -> ('accum, 'final) Base__Set_intf.Continue_or_stop.t) -> finish:('accum -> 'final) -> 'final

val fold_right : ('a, _) t -> init:'accum -> f:('a -> 'accum -> 'accum) -> 'accum

val iter2 : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> ('a, 'cmp) t -> f:([ `Left of 'a | `Right of 'a | `Both of 'a * 'a ] -> unit) -> unit

val filter : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> f:('a -> bool) -> ('a, 'cmp) t

val partition_tf : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> f:('a -> bool) -> ('a, 'cmp) t * ('a, 'cmp) t

val elements : ('a, _) t -> 'a list

val min_elt : ('a, _) t -> 'a option

val min_elt_exn : ('a, _) t -> 'a

val max_elt : ('a, _) t -> 'a option

val max_elt_exn : ('a, _) t -> 'a

val choose : ('a, _) t -> 'a option

val choose_exn : ('a, _) t -> 'a

val split : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> 'a -> ('a, 'cmp) t * 'a option * ('a, 'cmp) t

val group_by : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> equiv:('a -> 'a -> bool) -> ('a, 'cmp) t list

val find_exn : ('a, _) t -> f:('a -> bool) -> 'a

val nth : ('a, _) t -> int -> 'a option

val remove_index : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> int -> ('a, 'cmp) t

val to_tree : ('a, 'cmp) t -> ('a, 'cmp) tree

val to_sequence : comparator:('a, 'cmp) Comparator.t -> ?order:[ `Increasing | `Decreasing ] -> ?greater_or_equal_to:'a -> ?less_or_equal_to:'a -> ('a, 'cmp) t -> 'a Sequence.t

val binary_search : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> compare:('a -> 'key -> int) -> [ `Last_strictly_less_than | `Last_less_than_or_equal_to | `Last_equal_to | `First_equal_to | `First_greater_than_or_equal_to | `First_strictly_greater_than ] -> 'key -> 'a option

val binary_search_segmented : comparator:('a, 'cmp) Comparator.t -> ('a, 'cmp) t -> segment_of:('a -> [ `Left | `Right ]) -> [ `Last_on_left | `First_on_right ] -> 'a option

val merge_to_sequence : comparator:('a, 'cmp) Comparator.t -> ?order:[ `Increasing | `Decreasing ] -> ?greater_or_equal_to:'a -> ?less_or_equal_to:'a -> ('a, 'cmp) t -> ('a, 'cmp) t -> ('a, 'a) Sequence.Merge_with_duplicates_element.t Sequence.t