Module type Set.S_poly

type 'elt t

type 'elt tree

type comparator_witness

include Container.Generic with type ('a, 'b) t := 'a t with type 'a elt := 'a

val length : _ t -> int

val is_empty : _ t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'acc -> f:('acc -> 'a -> 'acc) -> 'acc

val fold_result : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'e) Result.t) -> ('acc, 'e) Result.t

val exists : 'a t -> f:('a -> bool) -> bool

val for_all : 'a t -> f:('a -> bool) -> bool

val count : 'a t -> f:('a -> bool) -> int

val sum : (module Container.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

val find : 'a t -> f:('a -> bool) -> 'a option

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val invariants : 'a t -> bool

val mem : 'a t -> 'a -> bool

override Container's mem

val add : 'a t -> 'a -> 'a t

val remove : 'a t -> 'a -> 'a t

val union : 'a t -> 'a t -> 'a t

val inter : 'a t -> 'a t -> 'a t

val diff : 'a t -> 'a t -> 'a t

val symmetric_diff : 'a t -> 'a t -> ('a, 'a) Either.t Sequence.t

val compare_direct : 'a t -> 'a t -> int

val equal : 'a t -> 'a t -> bool

val is_subset : 'a t -> of_:'a t -> bool

val are_disjoint : 'a t -> 'a t -> bool

module Named : sig ... end

val fold_until : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'final) Container.Continue_or_stop.t) -> finish:('acc -> 'final) -> 'final

val fold_right : 'a t -> init:'acc -> f:('a -> 'acc -> 'acc) -> 'acc

val iter2 : 'a t -> 'a t -> f:([ `Left of 'a | `Right of 'a | `Both of 'a * 'a ] -> unit) -> unit

val filter : 'a t -> f:('a -> bool) -> 'a t

val partition_tf : 'a t -> f:('a -> bool) -> 'a t * 'a t

val elements : 'a t -> 'a list

val min_elt : 'a t -> 'a option

val min_elt_exn : 'a t -> 'a

val max_elt : 'a t -> 'a option

val max_elt_exn : 'a t -> 'a

val choose : 'a t -> 'a option

val choose_exn : 'a t -> 'a

val split : 'a t -> 'a -> 'a t * 'a option * 'a t

val split_le_gt : 'a t -> 'a -> 'a t * 'a t

val split_lt_ge : 'a t -> 'a -> 'a t * 'a t

val group_by : 'a t -> equiv:('a -> 'a -> bool) -> 'a t list

val find_exn : 'a t -> f:('a -> bool) -> 'a

val nth : 'a t -> int -> 'a option

val remove_index : 'a t -> int -> 'a t

val to_tree : 'a t -> 'a tree

val to_sequence : ?order:[ `Increasing | `Decreasing ] -> ?greater_or_equal_to:'a -> ?less_or_equal_to:'a -> 'a t -> 'a Sequence.t

val binary_search : 'a t -> compare:('a -> 'key -> int) -> Binary_searchable.Which_target_by_key.t -> 'key -> 'a option

val binary_search_segmented : 'a t -> segment_of:('a -> [ `Left | `Right ]) -> Binary_searchable.Which_target_by_segment.t -> 'a option

val merge_to_sequence : ?order:[ `Increasing | `Decreasing ] -> ?greater_or_equal_to:'a -> ?less_or_equal_to:'a -> 'a t -> 'a t -> ('a, 'a) Sequence.Merge_with_duplicates_element.t Sequence.t

type ('a, 'cmp) set

val empty : 'a t

val singleton : 'a -> 'a t

val union_list : 'a t list -> 'a t

val of_list : 'a list -> 'a t

val of_sequence : 'a Sequence.t -> 'a t

val of_array : 'a array -> 'a t

val of_sorted_array : 'a array -> 'a t Or_error.t

val of_sorted_array_unchecked : 'a array -> 'a t

val of_increasing_iterator_unchecked : len:int -> f:(int -> 'a) -> 'a t

val stable_dedup_list : 'a list -> 'a list

val map : ('a, _) set -> f:('a -> 'b) -> 'b t

The types of map and filter_map are subtle. The input set, ('a, _) set, reflects the fact that these functions take a set of *any* type, with any comparator, while the output set, ('b, 'cmp) t, reflects that the output set has the particular 'cmp of the creation function. The comparator can come in one of three ways, depending on which set module is used

Set.map -- comparator comes as an argument

Set.Poly.map -- comparator is polymorphic comparison

Foo.Set.map -- comparator is Foo.comparator

val filter_map : ('a, _) set -> f:('a -> 'b option) -> 'b t

val of_tree : 'a tree -> 'a t