Module Base.Source_code_position

One typically obtains a Source_code_position.t using a [%here] expression, which is implemented by the ppx_here preprocessor.

type t = Lexing.position = {

pos_fname : string;

pos_lnum : int;

pos_bol : int;

pos_cnum : int;

}

See INRIA's OCaml documentation for a description of these fields.

sexp_of_t uses the form "FILE:LINE:COL", and does not have a corresponding of_sexp.

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

val sexp_of_t : t -> Sexplib0.Sexp.t

include Comparable.S with type t := t

include Comparisons.S with type t := t

include Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Or_error.t

include Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

val equal__local : t -> t -> bool

val compare__local : t -> t -> int

val to_string : t -> string

to_string t converts t to the form "FILE:LINE:COL".

val of_pos : (string * int * int * int) -> t

of_pos Stdlib.__POS__ is like [%here] but without using ppx.