Module Base.Stack

module type S = sig ... end

include S

type 'a t

include Sexplib0.Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

val t_sexp_grammar : 'a Sexplib0.Sexp_grammar.t -> 'a t Sexplib0.Sexp_grammar.t

include Invariant.S1 with type 'a t := 'a t

val invariant : ('a -> unit) -> 'a t -> unit

fold, iter, find, and find_map visit the elements in order from the top of the stack to the bottom. to_list and to_array return the elements in order from the top of the stack to the bottom.

Iteration functions (iter, fold, etc.) have unspecified behavior (although they should still be memory-safe) when the stack is mutated while they are running (e.g. by having the passed-in function call push or pop on the stack).

include Container.S1 with type 'a t := 'a t

val mem : 'a t -> 'a -> equal:('a -> 'a -> bool) -> bool

Checks whether the provided element is there, using equal.

val length : 'a t -> int

val is_empty : 'a t -> bool

val iter : 'a t -> f:('a -> unit) -> unit

val fold : 'a t -> init:'acc -> f:('acc -> 'a -> 'acc) -> 'acc

fold t ~init ~f returns f (... f (f (f init e1) e2) e3 ...) en, where e1..en are the elements of t

val fold_result : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'e) Result.t) -> ('acc, 'e) Result.t

fold_result t ~init ~f is a short-circuiting version of fold that runs in the Result monad. If f returns an Error _, that value is returned without any additional invocations of f.

val fold_until : 'a t -> init:'acc -> f:('acc -> 'a -> ('acc, 'final) Container.Continue_or_stop.t) -> finish:('acc -> 'final) -> 'final

fold_until t ~init ~f ~finish is a short-circuiting version of fold. If f returns Stop _ the computation ceases and results in that value. If f returns Continue _, the fold will proceed. If f never returns Stop _, the final result is computed by finish.

Example:

type maybe_negative =
  | Found_negative of int
  | All_nonnegative of { sum : int }

(** [first_neg_or_sum list] returns the first negative number in [list], if any,
    otherwise returns the sum of the list. *)
let first_neg_or_sum =
  List.fold_until ~init:0
    ~f:(fun sum x ->
      if x < 0
      then Stop (Found_negative x)
      else Continue (sum + x))
    ~finish:(fun sum -> All_nonnegative { sum })
;;

let x = first_neg_or_sum [1; 2; 3; 4; 5]
val x : maybe_negative = All_nonnegative {sum = 15}

let y = first_neg_or_sum [1; 2; -3; 4; 5]
val y : maybe_negative = Found_negative -3

val exists : 'a t -> f:('a -> bool) -> bool

Returns true if and only if there exists an element for which the provided function evaluates to true. This is a short-circuiting operation.

val for_all : 'a t -> f:('a -> bool) -> bool

Returns true if and only if the provided function evaluates to true for all elements. This is a short-circuiting operation.

val count : 'a t -> f:('a -> bool) -> int

Returns the number of elements for which the provided function evaluates to true.

val sum : (module Container.Summable with type t = 'sum) -> 'a t -> f:('a -> 'sum) -> 'sum

Returns the sum of f i for all i in the container.

val find : 'a t -> f:('a -> bool) -> 'a option

Returns as an option the first element for which f evaluates to true.

val find_map : 'a t -> f:('a -> 'b option) -> 'b option

Returns the first evaluation of f that returns Some, and returns None if there is no such element.

val to_list : 'a t -> 'a list

val to_array : 'a t -> 'a array

val min_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

Returns a minimum (resp maximum) element from the collection using the provided compare function, or None if the collection is empty. In case of a tie, the first element encountered while traversing the collection is returned. The implementation uses fold so it has the same complexity as fold.

val max_elt : 'a t -> compare:('a -> 'a -> int) -> 'a option

val of_list : 'a list -> 'a t

of_list l returns a stack whose top is the first element of l and bottom is the last element of l.

val create : unit -> _ t

create () returns an empty stack.

val singleton : 'a -> 'a t

singleton a creates a new stack containing only a.

val push : 'a t -> 'a -> unit

push t a adds a to the top of stack t.

val pop : 'a t -> 'a option

pop t removes and returns the top element of t as Some a, or returns None if t is empty.

val pop_exn : 'a t -> 'a

val top : 'a t -> 'a option

top t returns Some a, where a is the top of t, unless is_empty t, in which case top returns None.

val top_exn : 'a t -> 'a

val clear : _ t -> unit

clear t discards all elements from t.

val copy : 'a t -> 'a t

copy t returns a copy of t.

val until_empty : 'a t -> ('a -> unit) -> unit

until_empty t f repeatedly pops an element a off of t and runs f a, until t becomes empty. It is fine if f adds more elements to t, in which case the most-recently-added element will be processed next.

val filter_map : 'a t -> f:('a -> 'b option) -> 'b t

filter_map t ~f creates a new stack with only the elements for which f returns Some

val filter : 'a t -> f:('a -> bool) -> 'a t

filter t ~f creates a new stack with only the elements that satisfy f.

val filter_inplace : 'a t -> f:('a -> bool) -> unit

filter_inplace t ~f removes all elements of t that don't satisfy f.

val capacity : _ t -> int

capacity t returns the length of the array backing t.

val set_capacity : _ t -> int -> unit

set_capacity t capacity sets the length of the array backing t to max capacity (length t). To shrink as much as possible, do set_capacity t 0.