Module Type_equal.Id

Id provides identifiers for types, and the ability to test (via Id.same) at runtime if two identifiers are equal, and if so to get a proof of equality of their types. Unlike values of type Type_equal.t, values of type Id.t do have semantic content and must have a nontrivial runtime representation.

type 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

module type Arg0 = sig ... end

module type Arg1 = sig ... end

module type Arg2 = sig ... end

module type Arg3 = sig ... end

module type S0 = sig ... end

module type S1 = sig ... end

module type S2 = sig ... end

module type S3 = sig ... end

module Uid : sig ... end

Every Id.t contains a unique id that is distinct from the Uid.t in any other Id.t.

val uid : _ t -> Uid.t

val create : name:string -> ('a -> Sexp.t) -> 'a t

create ~name defines a new type identity. Two calls to create will result in two distinct identifiers, even for the same arguments with the same type. If the type 'a doesn't support sexp conversion, then a good practice is to have the converter be [%sexp_of: _], (or sexp_of_opaque, if not using ppx_sexp_conv).

Accessors

val hash : _ t -> int

val name : _ t -> string

val to_sexp : 'a t -> 'a -> Sexp.t

val hash_fold_t : Hash.state -> _ t -> Hash.state

same_witness t1 t2 and same_witness_exn t1 t2 return a type equality proof iff the two identifiers are the same (i.e., physically equal, resulting from the same call to create). This is a useful way to achieve a sort of dynamic typing. same_witness does not allocate a Some every time it is called.

same t1 t2 = is_some (same_witness t1 t2).

val same : _ t -> _ t -> bool

val same_witness : 'a t -> 'b t -> ('a, 'b) equal option

val same_witness_exn : 'a t -> 'b t -> ('a, 'b) equal

module Create0 (T : Arg0) : S0 with type t := T.t

module Create1 (T : Arg1) : S1 with type 'a t := 'a T.t

module Create2 (T : Arg2) : S2 with type ('a, 'b) t := ('a, 'b) T.t

module Create3 (T : Arg3) : S3 with type ('a, 'b, 'c) t := ('a, 'b, 'c) T.t