Module Base.Uchar

Unicode operations.

A Uchar.t represents a Unicode scalar value, which is the basic unit of Unicode.

See also String.Utf* submodules for Unicode support with multiple Uchar.t values encoded in a string.

type t = Uchar.t

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0.Sexp.t -> t

val sexp_of_t : t -> Sexplib0.Sexp.t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

type uchar := t

include Comparable.S with type t := t

include Comparisons.S with type t := t

include Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Or_error.t

include Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

val compare__local : t -> t -> int

val equal__local : t -> t -> bool

include Pretty_printer.S with type t := t

val pp : Formatter.t -> t -> unit

include Invariant.S with type t := t

val invariant : t -> unit

val succ : t -> t option

succ_exn t is the scalar value after t in the set of Unicode scalar values, and raises if t = max_value.

val succ_exn : t -> t

val pred : t -> t option

pred_exn t is the scalar value before t in the set of Unicode scalar values, and raises if t = min_value.

val pred_exn : t -> t

val is_char : t -> bool

is_char t is true iff n is in the latin-1 character set.

val to_char : t -> char option

to_char_exn t is t as a char if it is in the latin-1 character set, and raises otherwise.

val to_char_exn : t -> char

val of_char : char -> t

of_char c is c as a Unicode scalar value.

val int_is_scalar : int -> bool

int_is_scalar n is true iff n is an Unicode scalar value (i.e., in the ranges 0x0000...0xD7FF or 0xE000...0x10FFFF).

val of_scalar : int -> t option

of_scalar_exn n is n as a Unicode scalar value. Raises if not (int_is_scalar i).

val of_scalar_exn : int -> t

val to_scalar : t -> int

to_scalar t is t as an integer scalar value.

val utf_8_byte_length : t -> int

Number of bytes needed to represent t in UTF-8.

deprecated [since 2023-11] use [Utf8.byte_length]

val utf_16_byte_length : t -> int

Number of bytes needed to represent t in UTF-16.

deprecated [since 2023-11] use [Utf16le.byte_length] or [Utf16be.byte_length]

val min_value : t

val max_value : t

val byte_order_mark : t

U+FEFF, the byte order mark. https://en.wikipedia.org/wiki/Byte_order_mark

val replacement_char : t

U+FFFD, the Unicode replacement character. https://en.wikipedia.org/wiki/Specials_(Unicode_block)#Replacement_character

module Decode_result : sig ... end

Result of decoding a UTF codec that may contain invalid encodings.

module Utf8 : sig ... end

UTF-8 encoding. See Utf interface.

module Utf16le : sig ... end

UTF-16 little-endian encoding. See Utf interface.

module Utf16be : sig ... end

UTF-16 big-endian encoding. See Utf interface.

module Utf32le : sig ... end

UTF-32 little-endian encoding. See Utf interface.

module Utf32be : sig ... end

UTF-32 big-endian encoding. See Utf interface.

module type Utf = sig ... end