Module Uchar.Utf8

UTF-8 encoding. See Utf interface.

to_string encodes a Unicode scalar value in this encoding.

of_string interprets a string as one Unicode scalar value in this encoding, and raises if the string cannot be interpreted as such.

include Stringable.S with type t := Uchar.t

val of_string : string -> Uchar.t

val to_string : Uchar.t -> string

val byte_length : Uchar.t -> int

Returns the number of bytes used for a given scalar value in this encoding.

val codec_name : string

The name of this encoding scheme; e.g., "UTF-8".