Module Base__Binary_search

Functions for performing binary searches over ordered sequences given length and get functions.

These functions can be specialized and added to a data structure using the functors supplied in Binary_searchable and described in Binary_searchable_intf.

Examples

Below we assume that the functions get, length and compare are in scope:

(* Find the index of an element [e] in [t] *)
binary_search t ~get ~length ~compare `First_equal_to e;

(* Find the index where an element [e] should be inserted *)
binary_search t ~get ~length ~compare `First_greater_than_or_equal_to e;

(* Find the index in [t] where all elements to the left are less than [e] *)
binary_search_segmented t ~get ~length ~segment_of:(fun e' ->
  if compare e' e <= 0 then `Left else `Right) `First_on_right