Module Base__Bytes

OCaml's byte sequence type, semantically similar to a char array, but taking less space in memory.

A byte sequence is a mutable data structure that contains a fixed-length sequence of bytes (of type char). Each byte can be indexed in constant time for reading or writing.