Module Base__Ordering

Ordering is intended to make code that matches on the result of a comparison more concise and easier to read.

For example, instead of writing:

let r = compare x y in
if r < 0 then
  ...
else if r = 0 then
  ...
else
  ...

you could simply write:

match Ordering.of_int (compare x y) with
| Less -> ...
| Equal -> ...
| Greater -> ...