Module Base__Ref

Module for the type ref, mutable indirection cells r containing a value of type 'a, accessed with !r and set by r := a.

type 'a t = 'a Base__.Import.Caml.ref = {

mutable contents : 'a;

}

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val equal : ('a -> 'a -> bool) -> 'a t -> 'a t -> bool

include Base.Sexpable.S1 with type 'a t := 'a t

type 'a t

val t_of_sexp : (Base.Sexp.t -> 'a) -> Base.Sexp.t -> 'a t

val sexp_of_t : ('a -> Base.Sexp.t) -> 'a t -> Base.Sexp.t

val create : 'a -> 'a t

val (!) : 'a t -> 'a

val (:=) : 'a t -> 'a -> unit

val swap : 'a t -> 'a t -> unit

swap t1 t2 swaps the values in t1 and t2.

val replace : 'a t -> ('a -> 'a) -> unit

replace t f is t := f !t

val set_temporarily : 'a t -> 'a -> f:(unit -> 'b) -> 'b

set_temporarily t a ~f sets t to a, calls f (), and then restores t to its value prior to set_temporarily being called, whether f returns or raises.