Module Base__Sequence

A sequence of elements that can be produced one at a time, on demand, normally with no sharing.

The elements are computed on demand, possibly repeating work if they are demanded multiple times. A sequence can be built by unfolding from some initial state, which will in practice often be other containers.

Most functions constructing a sequence will not immediately compute any elements of the sequence. These functions will always return in O(1), but traversing the resulting sequence may be more expensive. The most they will do immediately is generate a new internal state and a new step function.

Functions that transform existing sequences sometimes have to reconstruct some suffix of the input sequence, even if it is unmodified. For example, calling drop 1 will return a sequence with a slightly larger state and whose elements all cost slightly more to traverse. Because this is sometimes undesirable (for example, applying drop 1 n times will cost O(n) per element traversed in the result), there are also more eager versions of many functions (whose names are suffixed with _eagerly) that do more work up front. A function has the _eagerly suffix iff it matches both of these conditions:

It might consume an element from an input t before returning.

It only returns a t (not paired with something else, not wrapped in an option, etc.). If it returns anything other than a t and it has at least one t input, it's probably demanding elements from the input t anyway.

Only *_exn functions can raise exceptions, except if the function underlying the sequence (the f passed to unfold) raises, in which case the exception will cascade.