Module Base

This module is the toplevel of the Base library; it's what you get when you write open Base.

The goal of Base is both to be a more complete standard library, with richer APIs, and to be more consistent in its design. For instance, in the standard library some things have modules and others don't; in Base, everything is a module.

Base extends some modules and data structures from the standard library, like Array, Buffer, Bytes, Char, Hashtbl, Int32, Int64, Lazy, List, Map, Nativeint, Printf, Random, Set, String, Sys, and Uchar. One key difference is that Base doesn't use exceptions as much as the standard library and instead makes heavy use of the Result type, as in:

type ('a,'b) result = Ok of 'a | Error of 'b 

Base also adds entirely new modules, most notably:

Comparable, Comparator, and Comparisons in lieu of polymorphic compare.

Container, which provides a consistent interface across container-like data structures (arrays, lists, strings).

Result, Error, and Or_error, supporting the or-error pattern.

The recommended way to use Base is to build with -open Base. Files compiled this way will have the environment described in this file as their initial environment.

module Applicative : sig ... end

module Array : sig ... end

module Avltree : sig ... end

module Backtrace : sig ... end

module Binary_search : sig ... end

module Binary_searchable : sig ... end

module Blit : sig ... end

module Bool : sig ... end

module Buffer : sig ... end

module Bytes : sig ... end

module Char : sig ... end

module Comparable : sig ... end

module Comparator : sig ... end

module Comparisons : sig ... end

module Container : sig ... end

module Either : sig ... end

module Equal : sig ... end

module Error : sig ... end

module Exn : sig ... end

module Field : sig ... end

module Float : sig ... end

module Floatable : sig ... end

module Fn : sig ... end

module Formatter : sig ... end

module Hash : sig ... end

module Hash_set : sig ... end

module Hashable : sig ... end

module Hasher : sig ... end

module Hashtbl : sig ... end

module Identifiable : sig ... end

module Indexed_container : sig ... end

module Info : sig ... end

module Int : sig ... end

module Int_conversions : sig ... end

module Int32 : sig ... end

module Int63 : sig ... end

module Int64 : sig ... end

module Intable : sig ... end

module Int_math : sig ... end

module Invariant : sig ... end

module Lazy : sig ... end

module List : sig ... end

module Map : sig ... end

module Maybe_bound : sig ... end

module Monad : sig ... end

module Nativeint : sig ... end

module Nothing : sig ... end

module Option : sig ... end

module Option_array : sig ... end

module Or_error : sig ... end

module Ordered_collection_common : sig ... end

module Ordering : sig ... end

module Poly : sig ... end

module Polymorphic_compare = Poly

module Popcount : sig ... end

module Pretty_printer : sig ... end

module Printf : sig ... end

module Linked_queue : sig ... end

module Queue : sig ... end

module Random : sig ... end

module Ref : sig ... end

module Result : sig ... end

module Sequence : sig ... end

module Set : sig ... end

module Sexpable : sig ... end

module Sign : sig ... end

module Sign_or_nan : sig ... end

module Source_code_position : sig ... end

module Stack : sig ... end

module Staged : sig ... end

module String : sig ... end

module Stringable : sig ... end

module Sys : sig ... end

module T : sig ... end

module Type_equal : sig ... end

module Uniform_array : sig ... end

module Unit : sig ... end

module Uchar : sig ... end

module Validate : sig ... end

module Variant : sig ... end

module With_return : sig ... end

module Word_size : sig ... end

include module type of struct include T end

module type T = sig ... end

module type T1 = sig ... end

module type T2 = sig ... end

module type T3 = sig ... end

module Sexp : sig ... end

module Export : sig ... end

include module type of struct include Export end

include module type of struct include Sexp.Private.Raw_grammar.Builtin end

val lazy_t_sexp_grammar : Base__Sexp.Private.Raw_grammar.t

type 'a array = 'a Array.t

val compare_array : a. ('a -> 'a -> int) -> 'a array -> 'a array -> int

val equal_array : a. ('a -> 'a -> bool) -> 'a array -> 'a array -> bool

val array_of_sexp : a. (Sexp.t -> 'a) -> Sexp.t -> 'a array

val sexp_of_array : a. ('a -> Sexp.t) -> 'a array -> Sexp.t

val array_sexp_grammar : Sexp.Private.Raw_grammar.t

type bool = Bool.t

val compare_bool : bool -> bool -> int

val equal_bool : bool -> bool -> bool

val hash_fold_bool : Hash.state -> bool -> Hash.state

val hash_bool : bool -> Hash.hash_value

val bool_of_sexp : Sexp.t -> bool

val sexp_of_bool : bool -> Sexp.t

val bool_sexp_grammar : Sexp.Private.Raw_grammar.t

type char = Char.t

val compare_char : char -> char -> int

val equal_char : char -> char -> bool

val hash_fold_char : Hash.state -> char -> Hash.state

val hash_char : char -> Hash.hash_value

val char_of_sexp : Sexp.t -> char

val sexp_of_char : char -> Sexp.t

val char_sexp_grammar : Sexp.Private.Raw_grammar.t

type exn = Exn.t

val sexp_of_exn : exn -> Sexp.t

type float = Float.t

val compare_float : float -> float -> int

val equal_float : float -> float -> bool

val hash_fold_float : Hash.state -> float -> Hash.state

val hash_float : float -> Hash.hash_value

val float_of_sexp : Sexp.t -> float

val sexp_of_float : float -> Sexp.t

val float_sexp_grammar : Sexp.Private.Raw_grammar.t

type int = Int.t

val compare_int : int -> int -> int

val equal_int : int -> int -> bool

val hash_fold_int : Hash.state -> int -> Hash.state

val hash_int : int -> Hash.hash_value

val int_of_sexp : Sexp.t -> int

val sexp_of_int : int -> Sexp.t

val int_sexp_grammar : Sexp.Private.Raw_grammar.t

type int32 = Int32.t

val compare_int32 : int32 -> int32 -> int

val equal_int32 : int32 -> int32 -> bool

val hash_fold_int32 : Hash.state -> int32 -> Hash.state

val hash_int32 : int32 -> Hash.hash_value

val int32_of_sexp : Sexp.t -> int32

val sexp_of_int32 : int32 -> Sexp.t

val int32_sexp_grammar : Sexp.Private.Raw_grammar.t

type int64 = Int64.t

val compare_int64 : int64 -> int64 -> int

val equal_int64 : int64 -> int64 -> bool

val hash_fold_int64 : Hash.state -> int64 -> Hash.state

val hash_int64 : int64 -> Hash.hash_value

val int64_of_sexp : Sexp.t -> int64

val sexp_of_int64 : int64 -> Sexp.t

val int64_sexp_grammar : Sexp.Private.Raw_grammar.t

type 'a list = 'a List.t

val compare_list : a. ('a -> 'a -> int) -> 'a list -> 'a list -> int

val equal_list : a. ('a -> 'a -> bool) -> 'a list -> 'a list -> bool

val hash_fold_list : a. (Hash.state -> 'a -> Hash.state) -> Hash.state -> 'a list -> Hash.state

val list_of_sexp : a. (Sexp.t -> 'a) -> Sexp.t -> 'a list

val sexp_of_list : a. ('a -> Sexp.t) -> 'a list -> Sexp.t

val list_sexp_grammar : Sexp.Private.Raw_grammar.t

type nativeint = Nativeint.t

val compare_nativeint : nativeint -> nativeint -> int

val equal_nativeint : nativeint -> nativeint -> bool

val hash_fold_nativeint : Hash.state -> nativeint -> Hash.state

val hash_nativeint : nativeint -> Hash.hash_value

val nativeint_of_sexp : Sexp.t -> nativeint

val sexp_of_nativeint : nativeint -> Sexp.t

val nativeint_sexp_grammar : Sexp.Private.Raw_grammar.t

type 'a option = 'a Option.t

val compare_option : a. ('a -> 'a -> int) -> 'a option -> 'a option -> int

val equal_option : a. ('a -> 'a -> bool) -> 'a option -> 'a option -> bool

val hash_fold_option : a. (Hash.state -> 'a -> Hash.state) -> Hash.state -> 'a option -> Hash.state

val option_of_sexp : a. (Sexp.t -> 'a) -> Sexp.t -> 'a option

val sexp_of_option : a. ('a -> Sexp.t) -> 'a option -> Sexp.t

val option_sexp_grammar : Sexp.Private.Raw_grammar.t

type 'a ref = 'a Ref.t

val compare_ref : a. ('a -> 'a -> int) -> 'a ref -> 'a ref -> int

val equal_ref : a. ('a -> 'a -> bool) -> 'a ref -> 'a ref -> bool

val ref_of_sexp : a. (Sexp.t -> 'a) -> Sexp.t -> 'a ref

val sexp_of_ref : a. ('a -> Sexp.t) -> 'a ref -> Sexp.t

val ref_sexp_grammar : Sexp.Private.Raw_grammar.t

type string = String.t

val compare_string : string -> string -> int

val equal_string : string -> string -> bool

val hash_fold_string : Hash.state -> string -> Hash.state

val hash_string : string -> Hash.hash_value

val string_of_sexp : Sexp.t -> string

val sexp_of_string : string -> Sexp.t

val string_sexp_grammar : Sexp.Private.Raw_grammar.t

type bytes = Bytes.t

val compare_bytes : bytes -> bytes -> int

val equal_bytes : bytes -> bytes -> bool

val bytes_of_sexp : Sexp.t -> bytes

val sexp_of_bytes : bytes -> Sexp.t

val bytes_sexp_grammar : Sexp.Private.Raw_grammar.t

type unit = Unit.t

val compare_unit : unit -> unit -> int

val equal_unit : unit -> unit -> bool

val hash_fold_unit : Hash.state -> unit -> Hash.state

val hash_unit : unit -> Hash.hash_value

val unit_of_sexp : Sexp.t -> unit

val sexp_of_unit : unit -> Sexp.t

val unit_sexp_grammar : Sexp.Private.Raw_grammar.t

Format stuff

type nonrec ('a, 'b, 'c) format = ('a, 'b, 'c) Stdlib.format

type nonrec ('a, 'b, 'c, 'd) format4 = ('a, 'b, 'c, 'd) Stdlib.format4

type nonrec ('a, 'b, 'c, 'd, 'e, 'f) format6 = ('a, 'b, 'c, 'd, 'e, 'f) Stdlib.format6

Sexp

Exporting the ad-hoc types that are recognized by ppx_sexp_* converters. sexp_array, sexp_list, and sexp_option allow a record field to be absent when converting from a sexp, and if absent, the field will take a default value of the appropriate type:

        sexp_array   [||]
        sexp_bool    false
        sexp_list    []
        sexp_option  None

sexp_opaque causes the conversion to sexp to produce the atom <opaque>.

For more documentation, see sexplib/README.md.

type 'a sexp_array = 'a array

type 'a sexp_list = 'a list

type 'a sexp_opaque = 'a

type 'a sexp_option = 'a option

List operators

include module type of struct include List.Infix end

val (@) : 'a list -> 'a list -> 'a list

Int operators and comparisons

include module type of struct include Int.O end

val (+) : int -> int -> int

val (-) : int -> int -> int

val (*) : int -> int -> int

val (/) : int -> int -> int

val (~-) : int -> int

val (**) : int -> int -> int

val abs : int -> int

val neg : int -> int

val zero : int

val (%) : int -> int -> int

val (/%) : int -> int -> int

val (//) : int -> int -> float

val (land) : int -> int -> int

val (lor) : int -> int -> int

val (lxor) : int -> int -> int

val lnot : int -> int

val (lsl) : int -> int -> int

val (asr) : int -> int -> int

val (lsr) : int -> int -> int

val (<) : int -> int -> bool

val (<=) : int -> int -> bool

val (<>) : int -> int -> bool

val (=) : int -> int -> bool

val (>) : int -> int -> bool

val (>=) : int -> int -> bool

val compare : int -> int -> int

val ascending : int -> int -> int

val descending : int -> int -> int

val equal : int -> int -> bool

val max : int -> int -> int

val min : int -> int -> int

Float operators

include module type of struct include Float.O_dot end

val (+.) : float -> float -> float

val (-.) : float -> float -> float

val (*.) : float -> float -> float

val (/.) : float -> float -> float

val (**.) : float -> float -> float

val (~-.) : float -> float

val (|>) : 'a -> ('a -> 'b) -> 'b

Reverse application operator. x |> g |> f is equivalent to f (g (x)).

val (@@) : ('a -> 'b) -> 'a -> 'b

Application operator. g @@ f @@ x is equivalent to g (f (x)).

Boolean operations

val (&&) : bool -> bool -> bool

val (||) : bool -> bool -> bool

val not : bool -> bool

val ignore : _ -> unit

val (^) : String.t -> String.t -> String.t

Common string operations

Reference operations

val (!) : 'a ref -> 'a

val ref : 'a -> 'a ref

val (:=) : 'a ref -> 'a -> unit

Pair operations

val fst : ('a * 'b) -> 'a

val snd : ('a * 'b) -> 'b

Exceptions stuff

val raise : exn -> _

val failwith : string -> 'a

val invalid_arg : string -> 'a

val raise_s : Sexp.t -> 'a

Misc

val phys_equal : 'a -> 'a -> bool

val force : 'a Lazy.t -> 'a

module Continue_or_stop : sig ... end

Continue_or_stop.t is used by the f argument to fold_until in order to indicate whether folding should continue, or stop early.

exception Not_found_s of Sexplib0.Sexp.t