Module Base.Char

A type for 8-bit characters.

type t = char

An alias for the type of characters.

val all : t list

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include Sexpable.S with type t := t

val t_of_sexp : Sexp.t -> t

val sexp_of_t : t -> Sexp.t

include Identifiable.S with type t := t

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include Sexpable.S with type t := t

val t_of_sexp : Sexp.t -> t

val sexp_of_t : t -> Sexp.t

include Stringable.S with type t := t

val of_string : string -> t

val to_string : t -> string

include Comparable.S with type t := t

include Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Or_error.t

include Comparator.S with type t := t

type comparator_witness

val comparator : (t, comparator_witness) Comparator.comparator

val validate_lbound : min:t Maybe_bound.t -> t Validate.check

val validate_ubound : max:t Maybe_bound.t -> t Validate.check

val validate_bound : min:t Maybe_bound.t -> max:t Maybe_bound.t -> t Validate.check

include Pretty_printer.S with type t := t

val pp : Formatter.t -> t -> unit

module O : Comparisons.Infix with type t := t

val to_int : t -> int

Returns the ASCII code of the argument.

val of_int : int -> t option

Returns the character with the given ASCII code or None is the argument is outside the range 0 to 255.

val of_int_exn : int -> t

Returns the character with the given ASCII code. Raises Failure if the argument is outside the range 0 to 255.

val unsafe_of_int : int -> t

val escaped : t -> string

Returns a string representing the given character, with special characters escaped following the lexical conventions of OCaml.

val lowercase : t -> t

Converts the given character to its equivalent lowercase character.

val uppercase : t -> t

Converts the given character to its equivalent uppercase character.

val is_digit : t -> bool

'0' - '9'

val is_lowercase : t -> bool

'a' - 'z'

val is_uppercase : t -> bool

'A' - 'Z'

val is_alpha : t -> bool

'a' - 'z' or 'A' - 'Z'

val is_alphanum : t -> bool

'a' - 'z' or 'A' - 'Z' or '0' - '9'

val is_print : t -> bool

' ' - '~'

val is_whitespace : t -> bool

' ' or '\t' or '\r' or '\n'

val get_digit : t -> int option

Returns Some i if is_digit c and None otherwise.

val get_digit_exn : t -> int

Returns i if is_digit c and raises Failure otherwise.

val min_value : t

val max_value : t