Module Float.O_dot

Similar to O, except that operators are suffixed with a dot, allowing one to have both int and float operators in scope simultaneously.

val (+.) : t -> t -> t

val (-.) : t -> t -> t

val (*.) : t -> t -> t

val (/.) : t -> t -> t

val (%.) : t -> t -> t

val (**.) : t -> t -> t

val (~-.) : t -> t