Module Base.Fn

Various combinators for functions.

val (|>) : 'a -> ('a -> 'b) -> 'b

A "pipe" operator.

val const : 'a -> _ -> 'a

Produces a function that just returns its first argument.

val ignore : _ -> unit

ignore is the same as Caml.ignore. It is useful to have here so that code that rebinds ignore can still refer to Fn.ignore.

val non : ('a -> bool) -> 'a -> bool

Negates a boolean function.

val forever : (unit -> unit) -> exn

forever f runs f () until it throws an exception and returns the exception. This function is useful for read_line loops, etc.

val apply_n_times : n:int -> ('a -> 'a) -> 'a -> 'a

apply_n_times ~n f x is the n-fold application of f to x.

val id : 'a -> 'a

The identity function.

See also: Sys.opaque_identity.

val compose : ('b -> 'c) -> ('a -> 'b) -> 'a -> 'c

compose f g x is f (g x).

val flip : ('a -> 'b -> 'c) -> 'b -> 'a -> 'c

Reverses the order of arguments for a binary function.