Module List.Or_unequal_lengths

Or_unequal_lengths is used for functions that take multiple lists and that only make sense if all the lists have the same length, e.g., iter2, map3. Such functions check the list lengths prior to doing anything else, and return Unequal_lengths if not all the lists have the same length.

type 'a t =

| Ok of 'a

| Unequal_lengths

val compare : ('a -> 'a -> int) -> 'a t -> 'a t -> int

val sexp_of_t : ('a -> Sexp.t) -> 'a t -> Sexp.t