Module Base.Option_array

'a Option_array.t is a compact representation of 'a option array: it avoids allocating heap objects representing Some x, usually representing them with x instead. It uses a special representation for None that's guaranteed to never collide with any representation of Some x.

type 'a t

include Sexpable.S1 with type 'a t := 'a t

val t_of_sexp : (Sexplib0.Sexp.t -> 'a) -> Sexplib0.Sexp.t -> 'a t

val sexp_of_t : ('a -> Sexplib0.Sexp.t) -> 'a t -> Sexplib0.Sexp.t

val empty : _ t

val create : len:int -> _ t

Initially filled with all None

val init_some : int -> f:(int -> 'a) -> 'a t

val init : int -> f:(int -> 'a option) -> 'a t

val length : _ t -> int

val get : 'a t -> int -> 'a option

get t i returns the element number i of array t, raising if i is outside the range 0 to length t - 1.

val get_some_exn : 'a t -> int -> 'a

Raises if the element number i is None.

val is_none : _ t -> int -> bool

is_none t i = Option.is_none (get t i)

val is_some : _ t -> int -> bool

is_some t i = Option.is_some (get t i)

These can cause arbitrary behavior when used for an out-of-bounds array access.

val unsafe_get : 'a t -> int -> 'a option

val unsafe_get_some_exn : 'a t -> int -> 'a

unsafe_get_some_exn t i is unsafe because it does not bounds check i. It does, however check whether the value at index i is none or some, and raises if it is none.

val unsafe_get_some_assuming_some : 'a t -> int -> 'a

unsafe_get_some_assuming_some t i is unsafe both because it does not bounds check i and because it does not check whether the value at index i is none or some, assuming that it is some.

val unsafe_is_some : _ t -> int -> bool

val set : 'a t -> int -> 'a option -> unit

set t i x modifies array t in place, replacing element number i with x, raising if i is outside the range 0 to length t - 1.

val set_some : 'a t -> int -> 'a -> unit

val set_none : _ t -> int -> unit

val swap : _ t -> int -> int -> unit

val clear : _ t -> unit

Replaces all the elements of the array with None.

Unsafe versions of set*. Can cause arbitrary behaviour when used for an out-of-bounds array access.

val unsafe_set : 'a t -> int -> 'a option -> unit

val unsafe_set_some : 'a t -> int -> 'a -> unit

val unsafe_set_none : _ t -> int -> unit

include Blit.S1 with type 'a t := 'a t

val blit : src:'a t -> src_pos:int -> dst:'a t -> dst_pos:int -> len:int -> unit

val blito : src:'a t -> ?src_pos:int -> ?src_len:int -> dst:'a t -> ?dst_pos:int -> unit -> unit

val unsafe_blit : src:'a t -> src_pos:int -> dst:'a t -> dst_pos:int -> len:int -> unit

val sub : 'a t -> pos:int -> len:int -> 'a t

val subo : ?pos:int -> ?len:int -> 'a t -> 'a t

val copy : 'a t -> 'a t

Makes a (shallow) copy of the array.