Module Base.Sexp

type t = Sexplib0.Sexp.t =

| Atom of string

| List of t list

Type of S-expressions

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include module type of Sexplib0.Sexp with type t := Sexplib0.Sexp.t

val t_of_sexp : Sexplib0.Sexp.t -> Sexplib0.Sexp.t

val sexp_of_t : Sexplib0.Sexp.t -> Sexplib0.Sexp.t

exception Not_found_s of Sexplib0.Sexp.t

exception Of_sexp_error of exn * Sexplib0.Sexp.t

val message : string -> (string * Sexplib0.Sexp.t) list -> Sexplib0.Sexp.t

val default_indent : int Stdlib.ref

val pp_hum : Stdlib.Format.formatter -> Sexplib0.Sexp.t -> unit

val pp_hum_indent : int -> Stdlib.Format.formatter -> Sexplib0.Sexp.t -> unit

val pp_mach : Stdlib.Format.formatter -> Sexplib0.Sexp.t -> unit

val pp : Stdlib.Format.formatter -> Sexplib0.Sexp.t -> unit

val to_string_hum : ?indent:int -> Sexplib0.Sexp.t -> string

val to_string_mach : Sexplib0.Sexp.t -> string

val to_string : Sexplib0.Sexp.t -> string

val of_float_style : [ `No_underscores | `Underscores ] Stdlib.ref

val of_int_style : [ `No_underscores | `Underscores ] Stdlib.ref

module Private : sig ... end

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

val invariant : t -> unit

val of_string : unit

Base has never had an of_string function. We expose a deprecated of_string here so that people can find it (e.g. with merlin), and learn what we recommend. This of_string has type unit because we don't want it to be accidentally used.

include Comparable.S with type t := t

include Comparisons.S with type t := t

include Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Or_error.t

include Comparator.S with type t := t

type comparator_witness

val comparator : ( t, comparator_witness ) Comparator.comparator