Module Base.T

This module defines various abstract interfaces that are convenient when one needs a module that matches a bare signature with just a type. This sometimes occurs in functor arguments and in interfaces.

module type T = sig ... end

module type T1 = sig ... end

module type T2 = sig ... end

module type T3 = sig ... end