Module Type_equal.Id

Id provides identifiers for types, and the ability to test (via Id.same) at runtime if two identifiers are equal, and if so to get a proof of equality of their types. Unlike values of type Type_equal.t, values of type Id.t do have semantic content and must have a nontrivial runtime representation.

type 'a t

val sexp_of_t : ('a -> Sexp.t) -> 'a t -> Sexp.t

module Uid : sig ... end

Every Id.t contains a unique id that is distinct from the Uid.t in any other Id.t.

val uid : _ t -> Uid.t

val create : name:string -> ('a -> Sexp.t) -> 'a t

create ~name defines a new type identity. Two calls to create will result in two distinct identifiers, even for the same arguments with the same type. If the type 'a doesn't support sexp conversion, then a good practice is to have the converter be <:sexp_of< _ >>, (or sexp_of_opaque, if not using pa_sexp).

val hash : _ t -> int

val name : _ t -> string

val to_sexp : 'a t -> 'a -> Sexp.t

val hash_fold_t : Hash.state -> _ t -> Hash.state

val same : _ t -> _ t -> bool

val same_witness : 'a t -> 'b t -> ('a, 'b) equal option

val same_witness_exn : 'a t -> 'b t -> ('a, 'b) equal