Module Base.Uchar

Unicode character operations.

A Uchar.t represents a Unicode code point -- that is, an integer identifying the character in abstract. This module does not provide any utilties for converting Uchar.ts to and from strings -- in order to do so, one needs to settle on a particular encoding, such as UTF-8 or UTF-16. See, for instance, the utf8_text library for converting to and from UTF-8.

type t = Base__.Import0.Caml.Uchar.t

val hash_fold_t : Hash.state -> t -> Hash.state

val hash : t -> Hash.hash_value

include Sexplib0.Sexpable.S with type t := t

val t_of_sexp : Sexplib0__.Sexp.t -> t

val sexp_of_t : t -> Sexplib0__.Sexp.t

val t_sexp_grammar : t Sexplib0.Sexp_grammar.t

include Comparable.S with type t := t

include Comparisons.S with type t := t

include Comparisons.Infix with type t := t

val (>=) : t -> t -> bool

val (<=) : t -> t -> bool

val (=) : t -> t -> bool

val (>) : t -> t -> bool

val (<) : t -> t -> bool

val (<>) : t -> t -> bool

val equal : t -> t -> bool

val compare : t -> t -> int

compare t1 t2 returns 0 if t1 is equal to t2, a negative integer if t1 is less than t2, and a positive integer if t1 is greater than t2.

val min : t -> t -> t

val max : t -> t -> t

val ascending : t -> t -> int

ascending is identical to compare. descending x y = ascending y x. These are intended to be mnemonic when used like List.sort ~compare:ascending and List.sort ~cmp:descending, since they cause the list to be sorted in ascending or descending order, respectively.

val descending : t -> t -> int

val between : t -> low:t -> high:t -> bool

between t ~low ~high means low <= t <= high

val clamp_exn : t -> min:t -> max:t -> t

clamp_exn t ~min ~max returns t', the closest value to t such that between t' ~low:min ~high:max is true.

Raises if not (min <= max).

val clamp : t -> min:t -> max:t -> t Or_error.t

include Comparator.S with type t := t

type comparator_witness

val comparator : ( t, comparator_witness ) Comparator.comparator

include Pretty_printer.S with type t := t

val pp : Formatter.t -> t -> unit

include Invariant.S with type t := t

val invariant : t -> unit

val succ : t -> t option

succ_exn t is the scalar value after t in the set of Unicode scalar values, and raises if t = max_value.

val succ_exn : t -> t

val pred : t -> t option

pred_exn t is the scalar value before t in the set of Unicode scalar values, and raises if t = min_value.

val pred_exn : t -> t

val is_char : t -> bool

is_char t is true iff n is in the latin-1 character set.

val to_char : t -> char option

to_char_exn t is t as a char if it is in the latin-1 character set, and raises otherwise.

val to_char_exn : t -> char

val of_char : char -> t

of_char c is c as a Unicode character.

val int_is_scalar : int -> bool

int_is_scalar n is true iff n is an Unicode scalar value (i.e., in the ranges 0x0000...0xD7FF or 0xE000...0x10FFFF).

val of_scalar : int -> t option

of_scalar_exn n is n as a Unicode character. Raises if not (int_is_scalar i).

val of_scalar_exn : int -> t

val to_scalar : t -> int

to_scalar t is t as an integer scalar value.

val utf8_byte_length : t -> int

utf8_byte_width t returns the number of bytes needed to represent t in the UTF-8 encoding (https://en.wikipedia.org/wiki/UTF-8).

val min_value : t

val max_value : t