Module Base.Word_size

For determining the word size that the program is using.

type t =

| W32

| W64

val sexp_of_t : t -> Sexp.t

val num_bits : t -> int

val word_size : t

Returns the word size of this program, not necessarily of the OS.