Module Base__Applicative_intf

Applicatives model computations in which values computed by subcomputations cannot affect what subsequent computations will take place.

Relative to monads, this restriction takes power away from the user of the interface and gives it to the implementation. In particular, because the structure of the entire computation is known, one can augment its definition with some description of that structure.

For more information, see:

      Applicative Programming with Effects.
      Conor McBride and Ross Paterson.
      Journal of Functional Programming 18:1 (2008), pages 1-13.
      http://staff.city.ac.uk/~ross/papers/Applicative.pdf

module type Basic = sig ... end

module type Basic_using_map2 = sig ... end

module type Applicative_infix = sig ... end

module type For_let_syntax = sig ... end

module type S = sig ... end

module type Let_syntax = sig ... end

module type Args = sig ... end

Argument lists and associated N-ary map and apply functions.

module type Basic2 = sig ... end

module type Basic2_using_map2 = sig ... end

module type S2 = sig ... end

module type Args2 = sig ... end

module type Applicative = sig ... end