Module Base__Backtrace

Module for managing stack backtraces.

The Backtrace module deals with two different kinds of backtraces:

Snapshots of the stack obtained on demand (Backtrace.get)

The stack frames unwound when an exception is raised (Backtrace.Exn)