Module Base__Or_error

Type for tracking errors in an Error.t. This is a specialization of the Result type, where the Error constructor carries an Error.t.

A common idiom is to wrap a function that is not implemented on all platforms, e.g.,

val do_something_linux_specific : (unit -> unit) Or_error.t