Module Base__Pretty_printer

A list of pretty printers for various types, for use in toplevels.

Pretty_printer has a string list ref with the names of pp functions matching the interface:

val pp : Format.formatter -> t -> unit

The names are actually OCaml identifier names, e.g., "Base.Int.pp". Code for building toplevels (this code is not in Base) evaluates the strings to yield the pretty printers and register them with the OCaml runtime.