Module Base__Validate

A module for organizing validations of data structures.

Allows standardized ways of checking for conditions, and keeps track of the location of errors by keeping a path to each error found. Thus, if you were validating the following datastructure:

{ foo = 3;
  bar = { snoo = 34.5;
          blue = Snoot -6; }
}

One might end up with an error with the error path:

bar.blue.Snoot : value -6 <= bound 0

By convention, the validations for a type defined in module M appear in module M, and have their name prefixed by validate_. E.g., Int.validate_positive.

Here's an example of how you would use validate with a record:

type t =
  { foo: int;
    bar: float;
  }
[@@deriving_inline fields][@@@end]

let validate t =
  let module V = Validate in
  let w check = V.field_folder t check in
  V.of_list
    (Fields.fold ~init:[]
       ~foo:(w Int.validate_positive)
       ~bar:(w Float.validate_non_negative)
    )

And here's an example of how you would use it with a variant type:

type t =
  | Foo of int
  | Bar of (float * int)
  | Snoo of Floogle.t

let validate = function
  | Foo i -> V.name "Foo" (Int.validate_positive i)
  | Bar p -> V.name "Bar" (V.pair
                             ~fst:Float.validate_positive
                             ~snd:Int.validate_non_negative p)
  | Snoo floogle -> V.name "Snoo" (Floogle.validate floogle)