Module Base_internalhash_types

type state = private int

state is defined as a subtype of int using the private keyword. This makes it an opaque type for most purposes, and tells the compiler that the type is immediate.

type seed = int

type hash_value = int

val create_seeded : seed -> state

val fold_int64 : state -> int64 -> state

val fold_int : state -> int -> state

val fold_float : state -> float -> state

val fold_string : state -> string -> state

val get_hash_value : state -> hash_value